Fix the logical failure in the testOwnerChangedBroadcast test case.
The runDeviceTests function uses the intrumentation command.
The instrumentation calls the forceStopPackageLocked function before starting the process.
The queued intent for this package in the BroadCastQueue is cleaned up
by forceStopPackageLocked function in approximately 1 second.
If a device can't deliver a DEVICE_OWNER_CHANGED within one second,
this test case always fail.
So, waiting for the broadcast idle state after sending a DEVICE_OWNER_CHANGED.
Signed-off-by: Seong kyu cho <sean.cho@lge.com>

Change-Id: I3b89cd16acd3731a8c6134cd431e41b9828b5eaf