Fix issue #28868955: CTS: PendingIntentTest failures; Regressed in NRD48x
These tests were a mess, having fixed sleeps all over the place waiting
for stuff to happen, causing the tests to take minutes to complete.

Fix these to just waiting for the expected action to occur.  Now the
entire suite runs in a few seconds.

Also fix a few of the tests, and expand the service test to also
test for getting results back.

Change-Id: Iefd92947b3af7d418341750ca24e0a6ec07edfd4