Fix TaskswitchingDeviceTest.
In the test body, the test sends an intent to activity and waits for a
reply from the resumed activity. However the test received the reply for
the intent sent by setUp method before the fix.

The CL explicitly waits for the reply in setUp method as well, so that
we don't receive a reply for the intent sent by setUp method.

BUG=28782775
Change-Id: I0a203bc66942a686ac7431ce36e6fc77d8c8c9fb