Modify findMinAndMax kernel to handle infinities in input.
The problem is that the accumulator function does not prefer the input
over the accumulator, and the combiner function does not prefer an
"accumulated" accumulator over an "initialized-only" accumulator.
This can cause the kernel to return an illegal location (-1, the
initialized location value) if the input consists of all
POSITIVE_INFINITY or all NEGATIVE_INFINITY.  (The easiest way to
provoke this would be with a single-cell input consisting of
POSITIVE_INFINITY.)

However, our current test cases never include infinities within the
input, so regardless of the pseudorandom seed used to construct the
input, they are safe.

Nevertheless, for the sake of clarity and completeness, it seems best
for the kernel to handle these cases.

Change-Id: I2067cb4c5766a80a7b7cad09f896ca938f896c68