Disable ListeningPortsTest
The test is originally introduced to help device manufacturers
to catch security vulnerabilities that would occur when they
ship by mistake a debugging daemon with open ports.

However the increased use of mDNS and other services triggered
a too high level of false negatives.

Hence the test is being disabled now, and will be reintroduced
later on when there is a better way to address the false
negatives.

Bug: 31803630
Change-Id: I044fa75ff13ed85e18eea4fcbb88fa22ce463b22