Add horizontal/vertical attributes for padding and margin
These new attributes provide a short-hand way for specifying the
separate left/right (or start/end) and top/bottom margin and padding.
It is common to use the same margin or padding on both edges, so these new
attributes make these common cases easier and less verbose.

Issue: 30261490 Better way to define paddingTop/Bottom/Left/Right in XML if vertical paddings are the same and horizontal paddings are the same
Test: This is the test
Change-Id: I07dedbebc62cd957ead22f3d22c35117d3e630e9