Fix Accessibility CTS: testFindByText in freeform mode
The test currently searches for all nodes within the
activity's view hierarchy, which includes the buttons
in the caption view in freeform mode and results in false
positives. Fix this only searching inside views that are
children of the test activity's main content view, just
like in verifyNodesInAppWindow.

Bug: 33621905
Test: CtsAccessibilityServiceTestCases pass in both
fullscreen mode and freeform mode.

Change-Id: I517e77737801cd58aee0cc6e3779471b17ab8d35