DO NOT MERGE - Disable ListeningPortsTest
The test is originally introduced to help device manufacturers
to catch security vulnerabilities that would occur when they
ship by mistake a debugging daemon with open ports.

However the increased use of mDNS and other services triggered
a too high level of false negatives.

Hence the test is being disabled now, and will be reintroduced
later on when there is a better way to address the false
negatives.

Cherry-picked from https://android-review.googlesource.com/#/c/293845/

Change-Id: I32c255e37f731467ef572a956afd840c0afbe347
Signed-off-by: Nobu Hayashi <nobuh@google.com>