Fix Loader bugs
Bug 32758881
Bug 34264461

When destroying the LoaderManager, the host was not being cleared.
The host was being leaked and this leaked the Activity.

When a LoaderManager is reused, for example with an orientation
change, the host was updated, but it wasn't started. This
adds a test for that case to ensure that when a reused
LoaderManager is loaded that it is properly started.

Test: ran fragment cts tests
Change-Id: Ib90f14e16083916a21a25ce87a4dd29ab9a33878