Remove failing UsesLibrary.testUsesLibrary test.
A fix for how duplicate classes were checked went in here:
https://android-review.googlesource.com/#/c/349268/

It exposed an issue with this CTS test. Now that the check works
properly, it's rejecting the test apk, which was declared to have
a shared library, but is compiled with no shared libraries when
installed. It doesn't seem like the runtime knows enough about this
library to pass it through to dex2oat.

There may be similar issues with the other tests in this file. A new
test should be developed to better check uses-library, but I'm removing
the failing test in the meantime.

Bug: 36285180
Test: cts-tradefed run singleCommand cts -d --module
CtsAppSecurityHostTestCases -t
android.appsecurity.cts.UsesLibraryHostTest

Change-Id: I4091e6da796f93df076c26f73f41a679c3dccd57