BuildTest: split out test expected to fail on prerelease builds
Build constants whose value is equal to CUR_DEVELOPMENT (10000)
are commonly used in development/prerelease builds of Android.
On release versions, this is not permitted.

This CL moves the stricter enforcement of values < 10000 into
a test method that is annotated
@android.platform.test.annotations.RestrictedBuildTest
to indicate that it is only suitable for releasable versions.

A slightly weaker check for values <= 10000 is run for all
builds of Android.

Bug: 35922665
Test: cts-tradefed run cts -m CtsOsTestCases -t android.os.cts.BuildTest
Change-Id: I999abe6c63a132776996ae955203553ab50d9622