cts: security: Generalize AslrTest
The AslrTest only the checks randomization of the stack memory, which is
fairly random even with ASLR disabled in the kernel. To improve the
test, also check the location of the shell binary, which does not change
much with ASLR disabled. The regular expression used to check the maps
file is further generalized to account for formatting differences
between kernels and devices.

Bug: 33351919
Test: Verified the test fails on kernels with ASLR disabled
Change-Id: Iebdb997e1066f80a19fd4fc3ee9cde9aa7c8e1af