[AWARE] Decouple tests from specifier Network Specifier API
The Network Specifier API is being updated from a String
to a custom class. This change removes a local variable storing the
created network specifier and passes it directly to the consuming
API. This allows test code to be built with either API - easing
merging flow.

Note: there are no functional changes in this CL.

Bug: 36282900
Test: builds
Change-Id: Ibf24ad4bff474ddfae62c6d5a28567abc9bfd6d7