[CTS] Add a test for key/value backup and restore.
This test is intended to replace BackupTest in CTS Verifier (removed in
this change).

The test installs the app that generates random values and saves them in
different shared preferences and files. It uses BackupAgentHelper to do
key/value backup of those values. The tests verifies that the values are
restored after the app is uninstalled and reinstalled.

Test: make cts -j40 && cts-tradefed run cts -m CtsBackupHostTestCases
Bug: 32797101
Change-Id: I75c0ebcd3bef3dae36a21ae704596af678c14ca9