Fix CtsBionicTestCases for translated devices
When TARGET_TRANSLATE_2ND_ARCH is true, both configured architectures
are 32-bit, so both of them would be installed into the lib32 directory.
Ideally we'd be using arch-specific directories instead of 32/64, but
this at least prevents the "overriding commands" warnings for translated
products, and ensures that we install the primary architecture's
libraries.

CtsBionicTestCases didn't compile under aosp_x86_arm before this change.

Test: lunch aosp_x86_arm-eng; m -j CtsBionicTestCases
Test: lunch aosp_arm64-eng; m -j CtsBionicTestCases (same before/after)
Change-Id: I8839a89160445eef14f261ac1396c6ec2387dec5