Fixes CTS AnimationBackgroundTest that would fail from race cond
This test would fail randomly based on the timing if the animation
finishes before the assert occurs such as if the launch activity builder
waits until the animation occurs. Fix this by increasing the duration of
the animation and only wait for the surface background to show before
asserting.

Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test
CtsServicesHostTestCases android.server.cts.
AnimationBackgroundTests#estAnimationBackground_duringAnimation
Fixes: 62219412

Change-Id: Idc824b9b2fc4a90878477547deeef4b5ec241a43