Fix AbsSeekBarTest
The test logic assumed that any change to a seekbar would cause a later
drawing operation on the view hierarchy. But bad layout parameters ensured
that the given seek bar had zero bounds, thus drawing and invalidation did
not work as expected.

The fix is to fix the layout attributes so that the seek bar actually has
real bounds, so invalidation and rendering should happen normally.

Test: manual, ran this test before/after fix
Bug: 62446922  CTS: 2 failures from AbsSeekBarTest; Process Crash
Change-Id: Ie1902a1eb3892fcbb9eaa5d8505add70836a2c8c