Fix wear CTS fail in testBroadcastNotReceivedForDifferentLauncher.
The session commit broadcast is sent to the "always" default launcher,
and if not available sent to the highest priority launcher.
Decreasing launcher priority will make it so that the test launcher
is never chosen as the "best" launcher.

Bug: 62272189

Change-Id: I31bb33147d334e3981f6c3b83de1ff7bc0298b43
(cherry picked from commit aff6ce05fc8bfd89a61d1428575ff7099c07751a)