Wait longer for device added notification.
Found that uhid driver takes a while to send out
device added notification in kernel 3.10. This was not
an issue in kernels 3.18+. The delay can be decreased
back to previous default value once 3.10 is no longer supported.
Also fixed the logic for checking whether a device added
notification has been received, since CountDownLatch does
not throw an exception upon timeout if the signal was
not received by the deadline.

Bug: 64981084
Test: bit
Change-Id: I56ee2ab51ac4bf809614160f0de64e2a88ff25d2
CtsHardwareTestCases:android.hardware.input.cts.tests.GamepadTestCase