Expanded testing for joystick actions.
Currently, only a single button on one joystick model
is tested in CTS. Here we expand the testing to include
additional KeyEvents and MotionEvents for Asus Gamepad
as well as the Sony DualShock 4. The eventual goal
is to add testing for all buttons and sticks for all
joysticks for which we have key layouts, thus allowing
us to ensure that a certain joystick model is supported
by all Android devices.

Rename the resource and the test files to follow a "company_model"
convention, so that multiple joystick models by the same company will
be shown next to each other (in alphabetic order).

Also, add an input library for cts tests. It will contain useful code
that may be utilized by different CTS test modules. Currently, it only
contains HidDevice for a simpler interface to use /dev/uhid.
It could also be used for testing shortcuts, for example (if it emulates
a keyboard instead of a joystick).

Test: atest AsusGamepadTestCase SonyDualshock4TestCase
Bug: 36069459
Bug: 111431828
Bug: 110270125
Change-Id: If0606b61feea9d4c293874d4c851b7611fbe2a9e