Reset activity and window variables when parsing protobuf
Protobuf3 has a feature that skips objects with unset, 0 or false values
and sets that object to null. When the test later tries to read from the
protobuf, some tests skip setting the variable back to 0, false or
null/empty because they check to see if the object is null and skip
setting a member variable causing the test to fail. This means that if
a test asserts toggling states (eg. check state is true, does something,
then check for false) it would fail if the check skips setting it back
to false when the object is null.

To fix this reset all states before parsing the protobuf.

Test: most existing tests
Change-Id: I0aa1d52fc3a27b66856e4b3f5489d4ca982cd04b
Fixes: 67650877