Explicitly set done to false upon test run failure
Previously we avoided setting done (to any value) upon test run failure, but
test runners like AndroidJUnitTest sometimes execute an additional test run
that does not fail, and mark the module done. To avoid flakiness, set done
to false when detecting a test run failure in which not all expected tests
are received.

bug: 68047779
Test: run cts-dev -m CtsPreferenceTestCases, disconnect/reconnect device mid-execution
Change-Id: I7013458bf0ec6aa009dbfb3a662150458ea1af8e