Enable IME CTS to test dummy input connection
In order to write regression tests for Bug 68841055, it would be useful
if we can test if InputMethodService#onStartInput() is called with a
real (non-dummy) InputConnection or not.

To do so, this CL reuses a CTS test
  InputMethodServiceDeviceTest#testOnStartInputCalledOnceIme1()
that was added for Bug 35599628 to see such an assertion is working as
intended.

Bug: 68841055
Bug: 69256929
Test: atest CtsInputMethodServiceHostTestCases
Change-Id: I3d62af2c52c0353b18d15fdbeccfeed59af5d05a