Account for vr virtual display
On a vr headset, the vr virtual display is always present and the 2d
activities are launched on it. Some of the tests currently assume any
activity launch happens on DEFAULT_DISPLAY. This change adds an
awareness of vr headset and tests the correct assumptions.

Bugs: 69859449, 69859653
Testing: Tests now pass.

Change-Id: Ieb999fd760679b48a836a3b430351f0e13221ab3
Signed-off-by: Karthik Ravi Shankar <karthikrs@google.com>