Fix too strict requirements in IME state handling
This is basically the hamcrest-free version of my previous attempt [1]
to relax the requirements when disabling/uninstalling the current IME.

We ended up removing the dependency on hamcrest, which is not yet
available in oreo-cts-dev branch.

Currently the following two tests essentially require the system to
always pick up the previously selected system IME when the CTS started
after disabling or uninstalling the current IME, which is actually
not guaranteed.

 - InputMethodServiceLifecycleTest#testDisableCurrentIme
 - InputMethodServiceLifecycleTest#testUninstallCurrentIme

It's true that the system tries to pick up an appropriate system IME
when the current one becomes unavailable, but which system IME will
be picked up is not deterministic.

This CL makes it clear and only requires that the new IME is not the
uninstalled/disabled one.

 [1]: I2b6563e9276e380dca485dbe5e3782f09e0dc445
      6cf3ccbf68865f7cde7fda5c0648588f3df60477

Bug: 69276863
Bug: 69893109
Bug: 70283683
Change-Id: I04160cabbc53a4a2b62b09242a0a125daa5f877b
Test: cts-tradefed run singleCommand cts -m CtsInputMethodServiceHostTestCases