Optimize window visibility check
This CL is a followup of I90306d4eba.

When WaitForValidActivityState specifies only #windowName and doesn't
specify #actvityName or #componentName as valid state for waiting, there is
no need for checking activity visibility.

This CL will shorten the testing time of ChildMovementTests and
DialogFrameTest to almost half (14 sec to 7 sec, and 4 min to 1.5 min
on modern phone).

Test: No affected CTS fails.
  $ cts-tradefed run singleCommand cts-dev \
        -m CtsWindowManagerDeviceTestCases \
	-t android.server.wm.ChildMovementTests
  $ cts-tradefed run singleCommand cts-dev \
        -m CtsWindowManagerDeviceTestCases \
	-t android.server.wm.DialogFrameTest
Bug: 70190274
Change-Id: I1a8f961d9ae239eb99f19863eb582862e3705260