DO NOT MERGE: Skip touch focus change on virtual displays for vr
Vr virtual display is used to render all 2d activities in vr. The
display is rendered in vr as a texture and has specific UX. Clicking on
anywhere not the virtual display on the screen does not dismiss the
virtual display. There is a dedicated 'BACK' button for this purpose.

Bug: 70336437
Test: testStackFocusSwitchOnTouchEvent passes

Change-Id: Ibae397655d1964f9b5a66e92ff2bd2e6f7f3bc3b
Signed-off-by: Karthik Ravi Shankar <karthikrs@google.com>