Update CTS test for new test launching behavior.
Previously, the way split-screen activities were launched was
modified to start the activity directly, rather than starting it
from the launching activity. This change meant that options passed
through the launching activity are no longer applied, such as
random data. For tests enforcing launching to side behavior, this
meant that the first TestActivity did not have random data for the
first launch intent. This caused a later assumption to be
incorrect.

This CL updates this code to instead launch later activities with
random intent data rather than none. This exercises the relevant
code paths in the same way as we are testing the filter comparison,
which does not have special behavior for null.

Change-Id: If56884624eff434fa23416727a6d746502cdc92a
Fixes: 70900940
Test: atest CtsActivityManagerDeviceTestCases:ActivityManagerSplitScreenTests