CTS test for statsd broadcastsubscription
Tests the generic broadcast subscription in statsd.
Sets a subscriber and makes sure it receives the broadcast.
Does this with multiple subscribers simultaneously.
Tests that removing the subscriber works properly.

Does not test SubscriberReporter's removeConfig since there seems to
be a bug in StatsManager.removeConfiguration which will prevent
it from working anyway.

Bug: 70356901
Test: cts-tradefed run cts-dev -m CtsStatsdHostTestCases -t android.cts.statsd.alert.BroadcastSubscriberTests
Change-Id: I1b4e688a511e44f845b3e038f4665f61d41dbe08