Restrict the minimum reported EARFCN to eliminate 0s
Some implementations are incorrectly reporting very
small numbers for the earfcn, especially zero. We cannot
declare that the channel number is invalid without cross
referencing other data; however, channels 0-6 are
actually not accessible by even the narrowest of
deployments. As a partial fix to at least flag a
problem, we can eliminate the first 7 channels from
the valid range.

At the same time, the appendix update to 25.101
clarifies that the minimum allowable channel for UTRAN
is 412 due to a 5Mhz channel width, so it is also
updated to properly restrict the minimum UARFCN.

Bug: 62535324
Test: CTS CellInfoTest passes on marlin
Merged-In: If7f18436bf278ca2703f2c373063ec73a9638510
Change-Id: If7f18436bf278ca2703f2c373063ec73a9638510