Allow expectEvent to wait for event start
Some IME events such as "showSoftInput" may trigger nested events such
as "onShowInputRequested" before the original "showSoftInput"
finishes.  When this happens, "onShowInputRequested" appears in the
event stream first followed by "showSoftInput".  This is because
we issue events when the event finishes, not when the event starts.

However, it is more common for us to write expectations about in what
order those events started rather than in what order those events
finished.  To make it easier to write tests, with this CL we starts
issuing two events ("enter" event and "exit" event) for each trace
action so that we can choose which signal the test is waiting for.

Although this is a non-trivial behavior change, existing CTS tests
should continue to pass (or fail if it's already broken) because so
far there is no IME CTS test that asserts strong ordering for events
that have different nest levels.

This CL also improves ImeEventStream#dump() to make it easier to see
how events are nested.  Here is a sample output.

  02-07 06:04:23.109  22243  onCreate: args={}
  02-07 06:04:23.116  22243  onCreateInputMethodInterface: args={}
  02-07 06:04:23.118  22243  attachToken: args=...
  02-07 06:04:23.122  22243  bindInput: args=...
  02-07 06:04:23.124  22243  onStartInput: args=...
  02-07 06:04:23.124  22243  [showSoftInput: args=...
  02-07 06:04:23.125  22243    [onShowInputRequested: args=...
  02-07 06:04:23.125  22243      onEvaluateInputViewShown: args={}
  02-07 06:04:23.126  22243      onEvaluateFullscreenMode: args={}
  02-07 06:04:23.126  22243    ]
  02-07 06:04:23.127  22243    onEvaluateFullscreenMode: args={}
  02-07 06:04:23.127  22243    onConfigureWindow: args=...
  02-07 06:04:23.128  22243    onEvaluateInputViewShown: args={}
  02-07 06:04:23.128  22243    onCreateInputView: args={}
  02-07 06:04:23.131  22243    onStartInputView: args=...
  02-07 06:04:23.141  22243  ]
  02-07 06:04:23.154  22243  onInputViewLayoutChanged: args=...
  02-07 06:04:23.279  22243  onInputViewLayoutChanged: args=...

Fix: 73077694
Test: atest CtsInputMethodTestCases
Test: prebuilts/checkstyle/checkstyle.py -f cts/tests/inputmethod
Change-Id: Ife7b70ca77a9ea72815e58238142e8cd170f856e