[Magnifier-22] Replace waitForIdleSync in CTS
In magnifier, we are now handling the pixel copy finish events on a
thread different from the UI thread. Therefore, it is not safe anymore
to assume that updates were applied when the main thread is idle.

This CL replaces waitForIdleSync with the usage of a callback called
from the magnifier when a requested operation has completed. For OEMs
this means that they also get more freedom to do work on background
threads.

Bug: 72041926
Test: atest CtsWidgetTestCases:android.widget.cts.MagnifierTest
Change-Id: I2f12e480019c0f78cc7daf62036986443f0edd5d