Build CtsVerifier without SDK
CtsVerifier is depending on other libraries which are not built with
SDK (e.g. telephony-common, bouncycastle, core-tests-support, etc.)

This is causing link-type check warning because dependency to the libs
actually indicates that CtsVerifier is using private APIs but the test
app itself is built with SDK (test_current). Since the link-type check
warning will be promoted to error soon and it seems that the
dependencies to the private APIs doesn't seem to be resolved anytime
soon, dropping LOCAL_SDK_VERSION from the cts app to be faithful to the
link-type check.

Bug: 69899800
Test: m -j CtsVerifier is successful and does not show any link-type
check warning

Change-Id: I25e3c9765f97d5f30033e045b5cca48c84ea99ae