Fix ActivityManagerPinnedStackTests#testEnterPipInterruptedCallbacks breaking states.
Before this change testEnterPictureInPictureDiscardSavedPositionOnFinish
is not able to launch PipActivity because somehow the activity in the
previous test is on the stack. I suspect that the am commands take a while
to finish while removeStacksWithActivityTypes thinks that it's already
in the middle of removing those activities from the stack. Changing the
order helps cleanup to be more effective.

Bug: 73905972
Test: atest
Change-Id: Ib4a20517c5f67a7a1aa7b3a01b3e9aea21f29261
CtsActivityManagerDeviceTestCases:ActivityManagerPinnedStackTests#testEnterPipInterruptedCallbacks,testEnterPictureInPictureDiscardSavedPositionOnFinish.