Fix a11y tooltip cts test
I got too cute and tried to make a single button both an
a11y pane (for the pane title tests) and a tooltip. That
results in the button promoting its content_change events
to window_state_change events, so they weren't showing up
in the wait for content changes.

Making one button a pane and a second one have a tooltip
eliminates the issue.

Also updating the focus test to keep it working.

Bug: 74599738
Test: atest AccessibilityEndToEndTest#testTooltipTextActionsReportedToAccessibility
Change-Id: I0636e478f6fca653a4cf5e7fd6ad4514c3d320d1