Do not use pause as an indicator to move forward in test.
It is possible that an activity is paused due to stack movement. For
example, Recents may gain focus during split screen. The secondary
stack's activity will enter pause at that point and  therefore won't
enter pause again when an activity is launched in the secondary
split screen.

This changelist removes this assumption and instead waits on an
activity in the secondary split screen to launch, which will
gaurantee focus has moved.

Change-Id: I85069d241ca1bbf1e0f01da81a5b5c4fdd0cd0c8
Fixes: 74989565
Test: atest android.server.am.lifecycle.ActivityLifecycleTests#testPausedWhenRecreatedFromInNonFocusedStack