Test EditText#setShowSoftInputOnFocus() with Mock IME
TextViewTest#testAccessShowSoftInputOnFocus() is currently testing the
behaviors of both the system and the currently selected IME, which is
known to be a source of non-determinism and test flakiness.  We should
use Mock IME to test only the system behavior.

Fortunately most of IME APIs that have been tested in
TextViewTest#testAccessShowSoftInputOnFocus() are already covered in
CtsInputMethodTestCases.

The only API we needed to port to CtsInputMethodTestCases is
EditText#setShowSoftInputOnFocus(), which is now coverted in
FocusHandlingTest with MockIME.

Fixes: 37570215
Test: make -j CtsWidgetTestCases
Test: atest CtsInputMethodTestCases
Change-Id: I3f56d8d0efbd3d61b5a97f39abc479c8ab568a1e