Launch recents placeholder when not present.
It is possible that the home activity will act as recents. If this
happens, the primary stack will be minimized in split screen. This
causes issues where we expect the primary stack to gain focus in
tests.

To resolve this issue, we launch an activity in place of recents.
This will allow the primary stack to unminimize.

Test: atest ActivityLifecycleTests#testPausedWhenRestartedFromInNonFocusedStack
Change-Id: I5addcd487367229a18704f3e3c447b0506ce1d75
Merged-In: I5addcd487367229a18704f3e3c447b0506ce1d75
Fixes: 77989482