Do not mark static libs as vts artifacts
This is not necessary as static libs are meant to be included into other
artifacts.

On the other hand any static lib marked as vts-artifacts is loaded into
the vts-tradfed process. In these two cases we are loading libraries meant
for on-device into a process on the host. This causes the mockito mock
maker to be configured for on-device behavior which fails.

Fixes: 80041014
Test: vts-tradefed run host --class com.android.tradefed.device.metric.VtsCoverageCollectorTest
      atest CtsActivityManagerDeviceTestCases
      atest CtsInputMethodTestCases
Change-Id: Ie1ee2dc2f9827a7038a9cad333c881d5b28cece2