Fix statsd cts test OverlayState
When run as standalone, this test passes consistently. But when run as
part of the suite, it sometimes fails on certain devices. This is
presumably due to some issue caused by the preceeding test. As a quick
fix, we can move this test to a more stable location (next to tests that
operate normally).

Bug: 79211560
Test: cts-tradefed run cts-dev -m CtsStatsdHostTestCases -t android.cts.statsd.atom.UidAtomTests
Change-Id: Ibc41047908c13f22a85d02d6bcb2eac0f8fa755c