Remove BatteryStats testWifiScan
This test was to ensure BatteryStats correctly recorded background wifi
scanning, since that is an important battery statistic due to its affect
on battery life. In P, background wifi scanning is throttled anyway, so
poor behaviour, in this regard, is no longer possible. Thus, the test is
no longer important. Since it fails due to the throttling, it may as
well be removed.

Test: removed
Bug: 79773672
Change-Id: Ic4a5f26152962338283602766a2cf2e84c1a4833