Set max mipmap level to prevent texture incompleteness.
The OpenGL ES 3.2 specification says that a texture is incomplete if it
has a minification filter that requires mipmaps and any of the mipmaps
between 0 and min(GL_TEXTURE_MAX_LEVEL, last mipmap level based on size)
are missing. Sampling an incomplete texture has undefined results, so
the test currently relies on undefined behavior of the Qualcomm driver
to pass. Set GL_TEXTURE_MAX_LEVEL to 0 if the texture does not have
mipmaps to fix this problem.

Bug: 110187458
Test: Built and ran the test on Pixel 2, passes.
Change-Id: Ie0bc70b5f5c95122cbedbdf1cc9e7e7eb0eaf6db