DO NOT MERGE: Change the way that find UI elements
If we change the style of notifications, a NullPointerException
may occur when looking up UI elements by the string of
'Permission requested' in a unit test.

So we change the way that find UI elements. Match UI elements
by means of regular expressions that starts with 'Permission requested'.

Test: run cts -m CtsSyncContentHostTestCases -t \
android.content.cts.SyncAdapterAccountAccessHostTest#\
testOtherCertAuthenticatorCanSeeAccount
Bug: 111417274

Change-Id: Ib044babbb09f993870fdac78954690acdb4d23cd