Fix flakes in various tests that depend on progressive animated values
The tests that check for a progression of values (testGetAnimatedValue(), etc)
are wrong. They set up an animator that is repeating, and then test values to ensure
that the values are always going up. But if the timing is off, even a little bit,
Then the object may be into a reversing cycle and the values at the end will
be going down.

Test: This is the test
Bug: 112667002
Change-Id: I6f401b2303b444364f27d3a5a2e81e8d6bc46092