DO NOT MERGE: Remove tests for injecting assistant key codes to reduce CTS flakiness.
The tests for injecting KEYCODE_ASSIST and KEYCODE_VOICE_ASSIST cause
inherent flakiness in the rest of the tests, because they cause the
assistant window to be launched, asynchronously, which ends up covering
other windows and causing them to lose focus (sometimes needed for
verification). The utility of ensuring that the key codes are not
received by the apps seems less important than the rest of the tests
being able to run deterministically.

Test: These are the tests
Bug: 112654616
Change-Id: I1176f2216468ed749986701f89a9c1515aa85f2c

cherry pick to pie-cts-dev
Bug: 111417080