Fix CTS GnssMeasurementWhenNoLocationTest test run failures
The CTS test android.location.cts.GnssMeasurementWhenNoLocationTest#
testGnssMeasurementWhenNoLocation fails occasionally because the location
is reported before the GNSS measurements are reported. One way to address
this issue is to disable network connections by enabling airplane mode
before running the test so that the GPS assistance data cannot be
downloaded and the measurements will get reported before the fix.

NOTE: When this same test is run through CtsVerifier, the airplane mode
      is not turned on because it was not possible to change the global
      setting airplane_mode_on when running as an application.

Bug: 111933267
Test: "adb shell am instrument -w -e class \
      android.location.cts.GnssMeasurementWhenNoLocationTest \
      android.location.cts/android.support.test.runner.AndroidJUnitRunner"
      Also, verified that the phone is put in airplane mode while the test
      is running. If the phone is already in airplane mode prior to the
      test, it is kept in that state.

Change-Id: I2826cf8b01aa3ab53605d8bf543592cffea3e3b4