min sdk needs to be less than split version
When aapt creates the configuration split, it removes any version
configuration that is before the APK's min sdk. Since we're splitting
on v7, the min sdk needs to be a value less than that.

If the min sdk is set, the build system will default the APKs to target
that sdk as well. So, need to manually set the target SDK to '27' to
allow the APKs to be installed as instant apps.

Bug: 113115970
Test: atest CtsAppSecurityHostTestCases:SplitTests
Change-Id: Id6c8f2577b5c47220f91b972bb5be18ef4f1ea2c