CtsNativeHardwareBuffer: Specify ES3 renderable type.
When choosing an EGL config, explicitly specify the OpenGL ES 3.x
renderable type. This fixes the test on some Mediatek devices, which don't
return the ES 3.x compatible config as the first one.

Additionally, don't try allocating a mipmapped AHardwareBuffer if the
extension required to bind it is not available.

Bug: 110732371
Test: externally by Mediatek
Change-Id: I48398490ed95c56fc30350a2658fa288c384f95a