Replace executeShell reboot by reboot() API
In order for the device APIs to be aware that a reboot
occured and how it should be handled, we need to use the
ITestDevice#reboot API (especially for remote device that
needs an extra connection step).

Since it's not the first time we fix a security test like
this, adding a forbidden list of commands.

Test: make cts
./cts-tradefed run cts-dev -m CtsSecurityBulletinHostTestCases
Bug: 114211519

Change-Id: Icf2b81c8ca106e4f5c7e0bd3d679cc3ed8ff0cca