Use average pixel distance to measure error
Use the average pixel distance instead of the L2 norm to measure the
error of the Rotation Vector Crosscheck's estimated test device
position.

The use of the L2 norm results in an error value that is difficult to
interpret. For instance, if there are three error values (3, 4, 5),
then the L2 norm produces an error of 7.07. If there is a fourth error
value of 1, the L2 norm produces a larger error of 7.14. Since the
error depends on the number of values being compared, the L2 norm is
not a good metric to determine the accuracy of the test device
position. Instead, a simple average should be used which is able to be
easily interpreted.

Also, update the asserts in Rotation Vector Crosscheck test to
correctly log the expected values if the test fails.

Bug: 112409976
Test: Verified test passes and fails as expected
Change-Id: I6d8ecdfcde392ef436bd5cdf284535501f67af52