AImageReaderCts: fix fence check logic
The original purpose of the test is to check whether
AImageReader_acquireNextImageAsync is generating either an legit fence
or -1 when no fence is needed. The test should check against "== 0"
rather than "!= kDummyFenceFd", as certain camera vendor may actually
implement the camera HAL in such a way that it yields a real fence.

Bug: 111313524
Test: cts-tradefed run commandAndExit cts --abi arm64-v8a -m CtsMediaTestCases -t android.media.cts.NativeImageReaderTest
Change-Id: Ibac2d0f438c1a182008b7d4a1ca3b25757940c8c