Increase wait time for AnimatedImageDrawable to end
Bug: 72737527
Bug: 120280954
Bug: 79156752
Test: atest
CtsGraphicsTestCases:android.graphics.drawable.cts.AnimatedImageDrawableTest

These tests are intended to verify that the the AnimatedImageDrawable
ends, but the if the device is running slowly, it may not have finished
in the exact expected time, or the message may not have been received
yet. Give the tests extra time before verifying that the animation has
ended.

Change-Id: Iedc97af2c9770cda4492c95a0e33313eef8cffad