Add CTS test for b/119189742
The test initializes an EditText and triggers selection on the text.
Before the bug was fixed in Iba5f6b606b1aaf2124487da5642c6f3b132a061c,
this would have crashed during the initialization of the
TextActionModeCallback instance, which was initializing the selection
handles when the handle drawables had not been loaded yet.

Test: atest android.widget.cts.TextViewTest\#testTextActionModeCallback_loadsHandleDrawables_b119189742
Bug: 119189742
Bug: 119214395
Change-Id: I1d010bb9047e44f6b1e89a4378ce5d4b2f4de31a