Compare pixel values in their premultiplied form.
Skia is updating our software rasterizer and as a result
has changed some pixel blending decisions.  This test compares
unpremultiplied values which doesn't account for the cases where
we now blend to a small alpha value with a fully saturated color
channel where the previous renderer rounded the alpha to zero.

Test: This is a test
Change-Id: I9b2ce0caeb77c4924aa834759c91eb02f1e16c4c