Add scrolling to isTextShown and tapOnViewWithText.
On a Pixel 2 phone,
android.preference2.cts.PreferenceActivityFlowLandscapeTest#listDialogLandscapeTest
passes because the needed UI elements barely render on-screen.

On the emulator, the screen resolution is a bit lower and scrolling is
needed to make the needed UI elements visible.

The logic here is ported from the pi-dev branch, where the same approach
was taken.

Bug: b/119130737
Test: Ran CTS tests on the emulator to assert that they now pass.
Change-Id: I0fd899973c90d2f5cff6051fab9872aea2033de5