Move a freeform WM test to simulated display.
Move testFreeformWindowManagementSupport to a large simulated display
when possible. Previouly the test was always run on primary display
which may not be able to provide a good user experience for freeform
windows so OEM may disallow implicit freeform window launching on it.
Move the test to a FHD simulated display and there shouldn't be any
reason for devices not to support freeform windows on it if it declares
support for freeform window management.

Bug: 119425708
Test: atest
ActivityManagerFreeformStackTests#testFreeformWindowManagementSupport

Change-Id: I20732fdc45f644a50907aa2f32264d9d50af5b5b