Add DefaultSmsAppHelper
This is used to allow the telephony provider CTS tests to change the
default SMS app, which is required for many operations against the
telephony providers.

Currently it uses the same logic as CtsProviderTestCases which grants
the WRITE_SMS appops permission to the test package, but does not
actually change the setting for default SMS app. That will be addressed
in a follow up commit.

Test: Added CTS tests covering the functionality, ran 100 times locally
Bug: 121044880
Change-Id: I3796cf476554679141035f7f6fbfb7de3f83fa7b