Skip testUsbAccessory for devices which don't have microphone or touchscreen.
According to CDD 7.1, devices which don't have microphone or touchscreen are not Handheld devices by section 2.1.
testUsbAccessory verifies 'if it is an Android Handheld device it MUST implement the AOA API.' in section 7.7.1, and should not be applied to non-Handheld devices.

Bug: 119172326
Change-Id: Ib2dec63c3a24d37bf5fcbaa964ba1e1276360293
Merged-In: Ib2dec63c3a24d37bf5fcbaa964ba1e1276360293