Add CheckLibraryNames
When the APIs are released, we need to update the make file for testing
new released APIs. To avoid doing this, I added the check-library-names
argument. When this argument is true, APIs test will be done if an API
library is available as the <uses-library> tag.

Bug: 123040759
Test: copy frameworks/base/test-base/api/current.txt into
prebuilts/sdk/28/public/api/android.test.base.txt and manually add some
APIs there.
run cts -m CtsSharedLibsApiSignatureTestCases

Change-Id: I6fa063e8c4d3450668dd73858ebfdda978f79f7c