Fix test failures caused by improvements in API signature files
1. Searches through class hierarchy to find methods.
2. Ignores final modifier on enum classes.

The API signature files (or at least the XML representation used by the
signature tests) were incomplete as they did not include all the methods
that were declared in non-API classes but which were exposed in the API
by the derived classes. The improvements to the generator fixed that
issue by pushing the methods down into the API classes to make it
appear as if they were declared in those classes.

Unfortunately, that change broke the test as the reflection code that
searched for the methods did not search through the through the class
hierarchy to find the methods. This change fixes that.

There was also a change in the output for enum classes which appeared
to mark all enum classes as being final even when they were not, e.g.
when an enum constant overrides a method and so creates a sub class.
As the compiler treats enums as effectively final even when they are
not this change does not check the final modifiers for enum classes.

Bug: 123254057
Test: runSignatureTests.sh CtsCurrentApiSignatureTestCases
Change-Id: Ifd60b1c72809dab64dd5416b9c1e3ffb3c8e7f42