Attempt to fix flaky ChangeBoundsTest.testInterruptWithReverseResizeClip
This CL changes the duration of the change bounds transition from
500ms to 1000ms so that there is a higher likelihood of catching
the transition in the middle of the transition. The theory here is
that the transition is caught after it passes the middle, but
by the time the reverse transition starts, the transition has
completed. By changing the duration to 1000ms, the reverse
transition should catch the animation while it is still in
the middle of animating.

Bug: 122739439
Test: ran ChangeBoundsTest several times on cuttlefish
Change-Id: I70f3a583441f11677ec23d8d2961009daca1223b