Restrict testVolume to STREAM_MUSIC on single volume devices
The test ensures the volume of ALARM, MUSIC, VOICE_CALL and RING streams
scales linearly when adjusting the volume (getVolumeDelta returns 1).

This does not apply to devices configured with single volume (eg. TV) as all
stream volumes are aliased to STREAM_MUSIC. When AudioManager rescales the
volume of the other streams, the volume step can be greater than 1 and test
will fail.

The change restricts the test to only operate on STREAM_MUSIC. This allows
to pass on a TV device configured with singleVolume && !fixedVolume.

Bug: 123656037
Change-Id: Id3ec17a3864183b5a31c09b142aa0b1e4d9d381b
Signed-off-by: Gregory Montoir <gregory.montoir@gmail.com>