Signature tests gather ancestor interfaces
This change fixes the bug that was revealed by I847066580248f47be609e2795db8904f58e550b8
in frameworks/base.

The change correctly added IInterface to the interfaces that a class
implementing an AIDL interface is implementing. But the api checker in
the signature test didn't gather IInterface from the runtime class.
As a result, the test thought the runtime class is not implementing an
interface that is documented.

This change fixes the problem by following parent(s) of interfaces so
that all interfaces that the class is implementing (either directly or
indirectly) are gathered.

Bug: 126374865
Test: atest CtsSystemApiSignatureTestCases:android.signature.cts.api.system.SignatureTest#testSignature
Change-Id: I9ae24d5d81a5e326ab81b453e79c2c4685f005b4