ActivityManagerProcessStateTest: wait for focus
ARC++ creates the windows in an asynchronous way. The window might not
be created, even though the activity is.

What this CL does, is to wait for activity to have focus and wait for possible
animations to finish after an activity is created.

This is needed to fix a flaky on ARC++ CTS tests.

Bug: 122971048
Test: atest -bit android.app.cts.ActivityManagerProcessStateTest
      Tested on Pixel Slate Chromebook
      Tested on Pixel One XL and Pixel Two phones
Change-Id: I30c175072782ea45f25fb007cd05611d052104fe