Configure signature test behavior for instant app/multi abi testing
None of the api signature tests (or the intent test) requires testing
on multiple ABIs as they do not use any native code.

The following tests should be run on instant apps to ensure that they
have access to the full API, intents and all the shared libraries that
an installed app has:
* CtsAndroidTestBaseCurrentApiSignatureTestCases
* CtsAndroidTestMockCurrentApiSignatureTestCases
* CtsAndroidTestRunnerCurrentApiSignatureTestCases
* CtsApacheHttpLegacyUsesLibraryApiSignatureTestCases
* CtsCurrentApiSignatureTestCases
* CtsIntentSignatureTestCases

The following tests should not be run as instant apps because doing so
provides no additional benefit over that provided by the above tests:
* CtsAndroidTestBase28ApiSignatureTestCases
* CtsApacheHttpLegacy27ApiSignatureTestCases
* CtsApacheHttpLegacyCurrentApiSignatureTestCases
* CtsSystemApiSignatureTestCases

Tried testing as per instructions in go/instant-cts-how but they did
not work, either in AOSP or git_master. Anytime that the following
was added to the cts-tradefed command it ran no modules.
    --enable-parameterized-modules --module-parameter instant_app

Uploaded so that it can be reviewed to see if there is something that
is wrong with these changes and also to trigger TH to test these.

Bug: 123604900
Bug: 127537194
Bug: 127538074
Bug: 123605076
Bug: 123660587
Bug: 123660829
Bug: 123605098
Bug: 123605059
Bug: 123660155
Bug: 123367456
Test: TH

Change-Id: Ie102ddb4e467e87e1531d164c3ca5fa6db3ff81e