APC: Reduce test flakiness with a bigger buffer
Because the capture will generate silence if the playback has not
received audio yet, the beginning of the recorded buffer can contain
silence even if the playback is playing.
Previously the buffer was ~200ms which in some rare case (~1/100) could
lead to no audio reaching the capture and the test failing.

As a result use a bigger buffer and stop as soon as audio is received.

Test: atest android.media.cts.AudioPlaybackCaptureTest#testCaptureMediaUsage
Bug: 111453086
Change-Id: Ic2c711ee05cd998bfe426db7e7b1e2bb10f6617e
Signed-off-by: Kevin Rocard <krocard@google.com>