aw: Fix testGetContentHeight() CTS test
This test was broken by change android:layout_height from wrap_content
to match_parent. [1] The viewport is larger than the actually content,
so getContentHeight() returns the height of viewport when match_parent.

This CL changes the pageHeight to a very large height so it will be
bigger than the viewport's height, so getContentHeight will return
the height of web content.

[1] https://android-review.googlesource.com/c/platform/cts/+/910193

Bug: 124826954, 129287096
Test: atest android.webkit.cts.WebViewTest#testGetContentHeight
Change-Id: I98b6bf97a1fcb8df9d93f29c64fae51a84081edd