add CTS to test the behavior of WindowInsets' 3 new APIs
To add tests to test the new APIs in WindowInsets. They are
getMandatorySystemGestureInsets, getTappableElementInsets,
getSystemGestureInsets.

The major focus is the interaction between the users and content
view. Sometimes it should trigger the system action and sometimes
it won't trigger the system action. WindowInsetsBehaviorTests focus
on the behavior of system guesture area.

These tests doesn't run on the devices that doesn't have system
gesture functionality.

Test: #in edge to edge mode
    atest -d android.systemui.cts.WindowInsetsBehaviorTests ;
Test: #2 button mode
    atest -d android.systemui.cts.WindowInsetsBehaviorTests ;
Test: #legcy navigation mode
    atest -d android.systemui.cts.WindowInsetsBehaviorTests ;
atest CtsSystemUiTestCases
Bug: 130346679

Change-Id: I2454235649b17c3c375dd28f3898e123b0e2cf35