Create "Using ANGLE" Toast Message Test
We've had cases where a change broke using the "ANGLE In Use" dialog box
because that functionality is not currently being tested.   This adds
tests that enables showing the dialog box and uses both ANGLE and the
native driver to verify neither leads to a crash or prevents ANGLE from
being loaded.

Bug: 129942954
Test: atest CtsAngleIntegrationHostTestCases
Change-Id: I607d8b2111cfd45828df32230dc627f781bf0b0c