Update WideColorGamutTests to be more flexible
Bug: 130753257
Test: This

WideColorGamutTests#testCanvasDrawColorLongBlendMode only passes because
"Wide Color Gamut" means P3. If we change it to F16 + EXTENDED_SRGB, the
test is incorrect. Update the test to be more flexible and support the
latter configuration.

Split the code that calls getPreferredWideGamutColorSpace into its own
method to share with testCanvasDrawColorLongBlendMode. When the
PreferredWideGamutColorSpace is null, fall back to SRGB instead of
EXTENDED_SRGB. The fallback happens on cuttlefish, when we are *not*
running the Activity in WideColorGamut mode, so the destination is SRGB.

Change-Id: I06e6a5765ed2fdab6ce68515ea5015283a325014