Update CtsAutoFillServiceTestCases to prevent test app in landscape mode will cause android.autofillservice.cts.ViewAttributesTest#checkViewLocationInAssistStructure failed.
Due to in landscape mode, screen height is not enough to show outerview to cause it’s height is not enough then test failed.
The patch is if outerview and EditText height is larger than parent rootcontainer height. It will reset EditText height to prevent outerView is out of the screen.

Test:
1. make CtsAutoFillServiceTestCases –j4
2. Replace CtsAutoFillServiceTestCases.apk in android-cts/testcases/
3. ./android-cts/tools/cts-tradefed
4. run cts -m CtsAutoFillServiceTestCases

Bug: 129384856
Change-Id: I00b845f916ae28102c926574417c5c44f0e9ec6a