Do not write sensor data to storage as Instant App
Instant Apps do not have access to external storage, and therefore,
the Sensor CTS tests cannot write data to external storage when CTS is
run in instant mode. This patch ensures that the tests do not try to
write to external storage whenever the tests are run in instant mode.

Bug: 131561181
Test: Verified Sensor CTS tests pass in instant mode with change, and
      fail without change
Test: Verified Sensor CTS tests pass in normal mode and save data to
      output directory

Change-Id: Ib1e346ab175f823495555fbac8c03bbb6be772a4