CTS test race condition on the memory puller atoms
testProcessMemoryState would start a test activity, trigger statsd
to pull the memory metrics and immediately force-stop the test app. This is
race condition which resulted either in pulling the test app memory while in
a zombie state (tests fail because rss = 0) or when killed
altogether (tests fail because the process is not found).

This CL ensures we do not force-stop the process immediately
(the same way as the rest on this class).

Test: UidAtomTests
Bug: b/137818141
Change-Id: Ie3d5f44e306fe2675f9b644c2c2f910d17edbc58