Fix shared uid permission test to match actual behavior.
Previously the test was passing APK path instead of package name to
PermissionManagerService so that it got the exception it expected. But
actually the behavior is the opposite if we fix that bug, and it's made
working this way deliberately in
BasePermission.enforceDeclaredUsedAndRuntimeOrDevelopment().

Exempt-From-Owner-Approval: The only owner is OOO.

Bug: 139047227
Test: atest SharedUidPermissionsTest
Change-Id: I4391930efc4c409dee9cef8849f73132cf1673d7
(cherry picked from commit 5a41cd43634c8964ac962cd60a5d01ba0d2f3d83)