Expand AndroidBitmap NDK testing
Bug: 135133301
Test: This

Add a test for all Bitmap.Configs, to verify that calling
AndroidBitmap_getInfo reports the correct AndroidBitmapFormat (including
the new RGBA_F16 format).

For HARDWARE, verify that the AndroidBitmapFormat matches the underlying
Config. Also verify that AndroidBitmap_lockPixels() fails. Rename
validateNdkAccessAfterRecycle to validateNdkAccessFails so that it can
be shared.

Add a test verifying that BAD_PARAMETER is returned for null jbitmap.

Change-Id: Id904c8cb4b22505635744ab271a639ace24ddbab