Add MANAGE_IPSEC_TUNNELS permission for System user and current user
The methods of class "IpSecManagerTunnelTest" under "CtsNetTestCases"
module fail with below error message:

"java.lang.SecurityException: android.net.cts from uid 1010126 not
allowed to perform MANAGE_IPSEC_TUNNELS"

In case of automotive there are two users, headless user which is the
background user, and there is a forground user as well. If any of the
two does not have the permission, the test case is failing. Hence
adding the permission for both the users.

Bug: 140392902
Test: cts-tradefed run cts -m CtsNetTestCases --test android.net.cts.IpSecManagerTunnelTest
Change-Id: If59f994445303749ec36a3bcecdc11365c2333bb
Signed-off-by: Aiswarya Cyriac <aiswarya.cyriac@intel.com>