Fix testTargetColorSpace w/o F16 HARDWARE support
Bug: 140284072
Bug: 140650033
Test: This

ImageDecoderTest#testTargetColorSpace verifies that when the client
requests an EXTENDED ColorSpace we promote the decode to F16. But if the
device does not support F16 + HARDWARE, we leave the decode at 8888, and
use the corresponding non-EXTENDED ColorSpace. Instead, use
ALLOCATOR_SOFTWARE to verify the original expectation.

In addition, add a test that verifies the HARDWARE + 8888 case described
above. Share code with BitmapTest for checking the native format.

Conflicts:
- tests/tests/graphics/src/android/graphics/cts/BitmapTest.java

Change-Id: I081d2cbc126239dd54fc9fe3d75a1978b09727ad