Exclude TYPE_HEART_RATE sensor type from CtsSensorTestCases[instant]
android.hardware.cts.SensorTest#testBatchAndFlushUidIdle TC requires android.permission.BODY_SENSORS permission.
According to the CTS document, some functions cannot be tested due to the permission problem of instant app, and these classes can be excluded from the test. https://source.android.com/compatibility/cts/cts-instant

The heartrate sensor is processed to be excluded from the test like above document in the code, but in reality, the test is performed in the instant mode and a fail occurs

Bug: b/141494456
Test: run cts -m CtsSensorTestCases -t  android.hardware.cts.SensorTest#testBatchAndFlushUidIdle

Change-Id: I482f4a3bd29e27726b8479fe401bf4f35144f488
Signed-off-by: Mark Hong <sungmin.h@samsung.com>