Add a timeout for unregistering a DAP
unregisterAudioPolicyAsync() is used when an AudioRecord.close(),
but the test expected a new policy to be registrable just after the
previous one was unregistered.

This leaded to a race condition if the unregisterAudioPolicyAsync did
not complete before the following line registerAudioPolicy.

This patch workaround the problem by retrying a few time, though it
might be better for R to make the close synchronous.

Test: AudioPlaybackCaptureTest
Bug: 141380721
Change-Id: I1721bc517308a131948804d800807c27985e456b
Signed-off-by: Kevin Rocard <krocard@google.com>