Reduce flakiness of multi-display tests
There is an assertion to ensure virtual/simulated
displays must be removed completely after closing.
By default it will wait 5 x 500ms for the removal,
but the testing device may not be fast enough and
then cause false alarm.

Especially for simulated display (it needs to wait
the overlay window is removed on system ui thread)
that is observed to take about 4s in the worst case.
Now change to 20 x 500ms to tolerate the slow cases.

Bug: 141674516
Test: atest MultiDisplayActivityLaunchTests
Change-Id: I3f5a57ad65c705cddd2b219dbf465940b5c2c905