Add output of "bmgr list transports" to exception msg when BackupPreparer fails
Failures to find local transport are happening again but the log added
in ag/9159255 gets lost because sponge only has the failure stacktrace
and doesn't have logcat for when our
preparer was run. Put the log into the exception message so that we can
see it in the stacktrace.

This is really redoing part of ag/9205305 again since that was reverted in ag/9225986

Bug: 143071114
Test: atest CtsBackupTestCases
Change-Id: Ic984a118381ab020bafb446393caa5a1e9f1db80