Abstract methods for instruction and button text.
The instruction for a test dynamically sets the button name.
For example in "Press the 'Start Test'" button the string
'Start Test' is dynamically set. This led to too long and complex
constructors of the test steps. That's why this change add abstract
method for getInstructionText() and getButtonStringId() in TestStepBase.

Bug: 142391087
Test: mma
Change-Id: I895198c5932e5c6370a2a9b8b6b46450440cf033