CtsDevicePolicyManagerTestCases not_secondary_user
Marks this CTS test module as NOT supporting secondary_user, meaning
that the tests are not eligible to be run from a regular Android user
other than the system user (i.e. other than user 0).

Note that 'user' here refers to the concept of users in Android
Multiuser; it is NOT the same as a uid.

Note that this module switches to system user in its
BaseDevicePolicyTest.setUp(); it cannot be run in secondary user and
is therefore being marked as such in this cl.

Bug: 141773058
Test: disabled for secondary user
Change-Id: I69c7cbe4c51c9bf6bffcd2138be9f3ca7f6728af