CTS secondary_user: CtsTfliteNnapiDelegateTestCases
Marks this CTS module as supporting secondary_user, meaning that
the tests are eligible to be run from a regular Android user other
than the system user (i.e. other than user 0).

Note that 'user' here refers to the concept of users in Android
Multiuser; it is NOT the same as a uid.

Bug: 141773058
Test: module passed with secondary_user params
 (--enable-optional-parameterization --module-parameter secondary_user)

Change-Id: Ifb2a6601ad73d1c96984c0af3ad2dbf3ccbef026