CTS GPS test
This test will enable the GPS and wait till we get a status or time to
fix message. This test was porting from already used CTS tests for
location. This has been edited to work in this application

Checks to see if our atom was filled out, meaning our write log code was
ran in GNSSMetrics

11/14 Update: Removed unneeded code from location CTS.

Test: atest UidAtomTests#testGpsStatus
Test: atest UidAtomTests#testGpsTimeToFirstFix
Bug: 141631388
Bug: 142740230
Change-Id: I9cbc8e41ec7fb68f7a64a0a71647a2c1abca1fde