CTS secondary_user: CtsOpenGlPerf2TestCases and CtsOpenGLTestCases
Marks these CTS modules as supporting secondary_user, meaning that
the tests are eligible to be run from a regular Android user other
than the system user (i.e. other than user 0).

Note that 'user' here refers to the concept of users in Android
Multiuser; it is NOT the same as a uid.

Bug: 141773058
Test: module passed with secondary_user params
 (--enable-optional-parameterization --module-parameter secondary_user)

Change-Id: Ica57e365faf3e6040c16d419df18bef9708032bb