Fix a race condition in ZenModeTest
We should wait for the real change of the ZEN_MODE/AIRPLANE_MODE_ON
settings instead of wait for the broadcast. Otherwise there can
be race condition, i.e., after the broadcast happened we check if
the value is changed to expected value, but the change may not happen
yet.

Bug: 143038550
Test: atest CtsVoiceSettingsTestCases
Change-Id: I69f31275ecb44fe05f9c1cd0c8d5b3eab5a3fd4b