Add tests to ensure session commit broadcast not sent
Test that SESSION_COMMITTED is not sent for certain staged install
flows. This should be tested before and after reboot, so that the
test will fail if the broadcast is sent before reboot (during
pre-reboot verification, for example).

To receive the broadcast, we need to register the test app as the
default home activity, which will send launcher broadcasts to the
test app.

Test: atest StagedInstallTest
Bug: 144270665
Change-Id: I2cfcfb1ea7f36681ea312db560d7d6fb39e03950