Enforce source for key events.
A lot of existing key event tests have wrong sources, so we need to fix
it. To make it easier as source is now a bit-wise flag in CTS tests, we
change string for unknown sources from "UNKNOWN" to empty string. That
also matches the value of InputDevice#SOURCE_UNKNOWN.

Bug: 145292999
Test: atest CtsHardwareTestCases
Change-Id: Ie2323f0a349b4d52030e767fe56eb6281f7b47b3