Relax gt150msDelta assert in testDaveyDrawFrame
The start time to calculate the total duration of a frame is its
frameInfo.intendedVsync. So there is a high possibility that if the
duration of the first frame is greater than 150_ms, then so is the second
frame (> 150_ms). This can make testDaveyDrawFrame fail sometimes.
Therefore we relax gt150msDelta assert in this case by increasing one
more upper threshold.

Bug: 145714153
Change-Id: I38938a895bfcdf10dd9c98fe30720266d69f7682
Test: run cts -m CtsIncidentHostTestCases -t com.android.server.cts
            .GraphicsStatsValidationTest#testDaveyDrawFrame