Destroy surface after collected test result for testReparentOff
Since the SurfaceControl is parented off-screen, 
if we release our reference it may completely die. 
If this occurs while the render thread is still 
rendering the RED background we could trigger a crash. 
For this test defer destroying the Surface until
we have collected our test results.

so, it's better to sleep some time after fill with color before
release surface control.

Bug: 146870018
Test:run cts -m CtsWindowManagerDeviceTestCases -t android.server.wm.SurfaceControlTest#testReparentOff
Change-Id: I5e54a9312db1138576c2a9f1f2138a7f613196a8