Allow unbuffered MotionEvent dispatch for View (2/2)
Add test cases to verify if motion events could be consumed as soon as
possible when the target view requested the unbuffered source.

The idea is injecting events as soon as possible and expect
the view would receive more than 1 events within 1 frame when it had
requested the unbuffered dispatch.

Bug: 135740001
Bug: 136277595
Test: atest ViewUnbufferedTest
Change-Id: I6dca85b0cfd51f08aac533656b3a65dc33860b6b