Check null before removing mHandlerThread.
If the setUp fails, mHandlerThread can be null. The test will clearly
fail anyway, but the NPE thrown here can mask the real issue because
tradefed will report the exception in the tearDown instead of the one
that happened during setUp().

If the setUpSucceeds, mHandlerThread will always be non-null. Ie. this
is a quality-of-life improvement for reporting failures, it doesn't
change the result of any of the tests.

Test: atest Camera2AndroidTestCase
Bug: 146994502
Change-Id: Ib8ef7fb79ab57c183708530e66e490ef6e86ea46