Verify that apps can insert subtitles and lyrics.
We don't have a specific API to insert these files through, so the
best option we have is the generic "Files" table.  However, that
table is typically pretty restrictive in where it allows things to
be placed.  To work around this, detect MEDIA_TYPE_SUBTITLE and
MEDIA_TYPE_PLAYLIST files and adjust our default/allowed values
accordingly.

Tests to verify that audio+lyrics and video+subtitles can be
inserted as sibling files in common directories.

Bug: 143707020
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: Ia6540b6ba575c3e3f6285fa5ea4be7b64b682a3f