Use a duplicate non-relaunched ShowWhenLockedActivity
Currently, we are relaunching the ShowWhenLockedActivity when
orientation change happens. That means, the window may not yet receive
window focus from input when we are trying to inject the back key.

The following sequence of events is observed:
1. Activity A is launched. It will eventually receive input focus (if
given enough time).
2. Rotation change happens.
3. Activity B is launched. Activity A no longer has focus.
4. Activity B will eventually receive input focus, if we wait long
enough.
5. Press BACK button
6. Activity B doesn't have focus yet, so the key gets dropped.
7. The test fails because Activity B is still shown over the keyguard.

To remove the race, we wait until the activity has focus. We also allow
the activity to handle orientation changes. This would prevent
Activity B from getting launched, and therefore will
ensure that Activity A has input focus. The new sequence
of events will be:
1. Activity A is launched.
2. Wait until Activity A has focus.
3. Rotation change happens.
4. Press BACK button.
5. Activity A receives BACK button, and gets dismissed.
6. Keyguard is now on top, test passes.

Also, we remove the flag FLAG_SHOW_WHEN_LOCKED from the window of the
original activity. That flag is deprecated. Instead, we add the
equivalent setting in the manifest.

Bug: 70668286
Test: atest android.server.wm.KeyguardTests#testUnoccludedRotationChange
Change-Id: I62cc4f4bb9c7deb67689aaeda506faa0ea2e29c2