CtsSimpleperfTestCases secondary_user
Marks this CTS test module as NOT supporting secondary_user, meaning that
the tests should not be run from a user other than the system user.

The module currently always switches to system user for execution - so we're
just labeling it as such using parametrization.

Note that 'user' here refers to the concept of users in Android
Multiuser; it is NOT the same as a uid.

Bug: 141773058
Test: module passed with secondary_user params
 (--enable-optional-parameterization --module-parameter secondary_user)

Change-Id: Ib545c0834a56367cad1b145e67848c5f0ab70fc9