Test AImageDecoder with dataspaces
Bug: 135133301
Bug: 139124386
Test: This

Test AImageDecoderHeaderInfo_getDataSpace and AImageDecoder_setDataSpace.

Add a dataspace parameter to nTestInfo, so we can verify the dataspace.

Test that the dataspace is ignored by A_8.

Add a test image that is LINEAR_EXTENDED_SRGB. testGetDataSpace verifies
that AImageDecoderHeaderInfo_getDataSpace properly reports all
ADataSpace values, except for the EXTENDED variants. (It uses
Bitmap#compress to test each value, but compress will never result in an
encoded image that defaults to EXTENDED on Android.) EXTENDED_SRGB is
already covered by an Asset, so this covers the final case.

Make decodeAgain return a boolean, so we can exit on failure, avoiding
throwing an Exception while one is pending.

Change-Id: Iffe659e50078139188c3325545624640ae177cc2