CtsSystemApiAnnotationTestCases accepts parameterized annotation
SystemApi annotation is parameterized with the client and the process
atttributes. Existing @SystemApis for privileged apps now become
@SystemApi(client=PRIVILEGED_APPS,process=ALL) and we have more
@SystemApis for modules with different client and process attribures.

However, the test currently doesn't understand the parameterized
annotation. It treats all @SystemApi annotations the same. As a result,
it fails because the new @SystemApis for modulese are also gathered by
the test but they don't appear in the API signature text file which is
only for the legacy @SystemApi.

Fixing the problem by allowing the test to accept an annotation spec
(e.g. @SystemApi(client=PRIVILEGEDE_APPS,process=ALL), instead of just
an annotation name.

Bug: 147812912
Test: atest CtsSystemApiAnnotationTestCases
This still fails for anther reason. API symbols from ike.jar (which is
not part of the boot class path) are still in the API signature file.

Change-Id: I62b3f4035456effaac136a1514aa80369b8b3a41