Specify the boundsToCheck to correspond to the testing content area only.
Prior to this change, boundsToCheck corresponded to the entire display
screen. However, even though we enter full-screen mode for this test,
based on configruation, some SystemUI components like NavigationBar or
StatusBar may be included in the boundsToCheck, which results in false
test failures in SurfaceViewSyncTest. BoundsToCheck should only include
the area that is relevant to the test.

Bug: 146536203
Test: SurfaceViewSyncTest now passing even when some SystemUI components
are shown on the screen.

Change-Id: I0f9e70d5ae3b1587420de2b33da17b269e60613d