RESTRICT AUTOMERGE
This change is the union of I9fb2756ef3257a9c393f5407060c83e3e567f99b
and Ib5a24f8be81bba658273e4a53ef63a8603dfa853 with the addition of a fix
for b/152342138

Original CL descriptions:

*** Ib5a24f8be81bba658273e4a53ef63a8603dfa853
Add test for Presentation

The helper activity tries to display a Presentation on the display given
by the "display_id" extra and the test verifies that the Presentation is
shown on Displays on public displays that has the presentation bit set
and verfies that it's blocked on other displays.

*** I9fb2756ef3257a9c393f5407060c83e3e567f99b
Don't explicitly set windowType for presentation

Presentation will automatically set the window type to
TYPE_PRESENTATION_PRIVATE if the Presentation is targeting a private
virtual display.
***

Bug: 141745510
Bug: 152342138
Test: atest CtsWindowManagerDeviceTestCases:android.server.wm.PresentationTest CtsDisplayTestCases:android.display.cts.VirtualDisplayTest

Change-Id: I165159d7bd1c59f27598a73b7f943ea87e97c957