Fail-fast hostside IME CTS tests upon system crash
As we have seen in Bug 150304313, it has always been painful to figure
out what is going on when hostside IME CTS tests are failing due to
random system restart bugs.  Thus here we start protecting ourselves
by showing more descriptive message when a system crash is detected
while running hostside IME CTS tests.

The basic idea is to rely on a volatile system property

  sys.system_server.start_count,

which was recently added [1] as part of our effort on Bug 124022170,
to fail-fast after the count has changed during test runs.

For instance, suppose the following test class.

  @RunWith(DeviceJUnit4ClassRunner.class)
  public class MyTestClass extends RebootFreeHostTestBase {
      @Test
      public void test1() throws Exception {
          // do something;
      }

      @Test
      public void test2() throws Exception {
          // do something;
      }
      @Test
      public void test3() throws Exception {
          // do something;
      }
  }

If the device soft-rebooted between test1() and test2(), then this
base class RebootFreeHostTestBase would let test2() and test3()
immediately fail by throwingIllegalStateException with saying that the
device reboot is detected and the last test method before reboot is
test1().

 [1]: Ie3e77903e723eb5b7e79b545d4287d9b07c4b379
      b4739d9d14d55ce6193f54b3cb248196facf947f

Bug: 150773958
Test: Manually verified as follows.
 1. Flash a build that is affected by Bug 150039791
 2. atest CtsInputMethodServiceHostTestCases --instant
 3. Confirm tests fail with IllegalStateException after system crash
Change-Id: I8c2cff5de1f2a7a3c8991fa09c77f80568ef79ea