Fix failed test case when test on TV
Description:
In the sleepDevice() funtion, "preeSleepButton" will be sent to TV and run on TV.
Actually, TV will go to sleep and can't receive power key event to wake up.
At this point, an exception of DeviceNotAvailableException is catched and the rest
of the test case will not go on. Finally, the test will fail.
we got the suggestion from goole to skip these test case when we test on TV.

Bug: 141446963
Test: cts test on eldridge p pass
Test: on TV, run cts -m CtsWindowManagerDeviceTestCases

Change-Id: I23a428d42738ebcd654891d4a881f1a714c4110f
Merged-In: I23a428d42738ebcd654891d4a881f1a714c4110f