Add tests for Bitmap#asShared
Bug: 150395371
Test: This

Verify that asShared returns a new Bitmap that is identical.
Share code with testNdkCompress to create a Bitmap with some interesting
content. Compress HARDWARE Bitmaps to compare properly, since we cannot
compare their pixels directly.

Verify that if the original Bitmap is already immutable and in shared
memory the method returns itself.

Change-Id: I442cd0c7a849e4c6a2fe7ba21b25a6e44417bbf5