Update Canvas#quickReject tests
Bug: 152323890
Test: This

This method had three overloads, each of which was recently replaced
with an overload that does not use the deprecated Canvas.EdgeType.

Update the existing test to drop the EdgeType. Test the other two
overloads, and add matrices to verify that the rejection is affected by
the matrix.

Change-Id: I1e4afefe1ab077a0c47c23b168b8c20cfea07ffa
(cherry picked from commit 57396415f49ae9f0db2236464d8eb8d348415ce2)