Bug fixed for CTS failures on CtsAppSecurityHostTestCases
On the small screen(ex: 480x854 with 320 density), the perference of
the "Storage & cache" does not show on the screen, the screen only show
the perference of the "Notifications" and "Permissions" on the App info
page. Use the UiScrollable function to find the "internal storage"
object to avoid this issue on the small screen.

Bug: 152539673
Change-Id: Ia0b240ebebf30b09027282b95e90caf0b117ba58
Test: m CtsAppSecurityHostTestCases