Fix errors in AccessibilityGestureDispatchTest
Gesture dispatch has been updated to be more granular for
AccessibilityService instances targeting R or later. The granularity is
determined by the refresh rate of the device. As such it can generate
much smaller slices in time between frames of the gesture. This has
resulted in AccessibilityGestureDispatchTest failing due to the inexact
nature of floating point representations.

Test: atest
Bug: 150190594
Change-Id: I680f8252ec83cc8a01e85c51fd582b0c8b7ba038