Add a test for grayscale color spaces
Bug: 153410582
Test: this

Add gimp-d65-grayscale.jpg. This image was created with GIMP, using its
built-in D65 grayscale color profile, which triggers the bug.

Verify that a Bitmap decoded from this image has a ColorSpace with the
properties we expect (all 1s in primaries, all 0s in transforms, except
the diagonal). Verify that when creating a new Bitmap with this same
ColorSpace (via createBitmap or createScaledBitmap) the resulting Bitmap
has an equal ColorSpace. Further, verify that the new Bitmap is not all
black.

Change-Id: I0a280cdfdb02766f5e415408386cce49a59504ab