Call IMM#showSoftInput on the UI thread in CTS
This is a small code clean up for our previous CL [1], which
introduced SearchViewTest#testShowImeWithSearchViewFocus.

Senamtically runOnMainSync() should have been used to call
InputMethodManager#showSoftInput() for the SearchView that is bound to
the UI thread.  So far we haven't seen any pariticular test failure
because of this cross-thread API call though.

As long as the test has been passing, there should be no developer
partner observable requirement changes.

 [1]: Iff700cc4db8e7560a103f993653d89545d835245
      761b2974a1205dfab44b978b5f33e3a31fe38fbe

Bug: 112399521
Bug: 124461829
Test: atest CtsInputMethodTestCases
Change-Id: I7a6af6fa5638243f403b3cf3a22e45b284faf34c