Let test IME behave the same even w/ hard keyboard
This is a preparation to introduce keyboard visibility assertions for
hostside IME CTS.  To make it possible, the software keyboard needs to
be shown no matter whether a hardware keyboard is attached or not,
which is often the case on emulator devices.

Note that this was already enforced to MockIME [1][2].

All the existing hostside IME tests are expected to continue passing.

 [1]: I3a108a9e776f62496ac0ab809536b1cec197e60f
      cfd2689ca09f0ea9d58b67434124604d83c4e3f6
 [2]: I463bb4e9ae665b93950d664e0704ca4e938d32a8
      91923927c19d51e07cdee7171c59c84626638762

Bug: 153454383
Bug: 153580136
Test: make -j CtsInputMethod1 \
              CtsInputMethod2 \
              CtsInputMethodServiceDeviceTests \
              CtsInputMethodServiceEventProvider \
              CtsInputMethodServiceHostTestCases \
              EditTextApp \
          && atest CtsInputMethodServiceHostTestCases
Change-Id: I45f7b2111904e5632a974bfa7a72e72fa9506014