Do not use UI thread for heavy computations
Currently, GenerateImagesActivity runs a heavy task, compressOutput, on
UI thread. This is problematic because the focus change events are now
coming from input, and also therefore also delivered on the UI thread.
If they don't get processed in a timely manner (currently, 5 seconds),
input sends ANR for this connection.

The order of events is roughly as follows:

Activity A starts Activity B and receives the result in
onActivityResult. When Activity B finishes, it loses focus, and
onActivityResult is called on UI thread.

Actitivy A starts doing heavy work inside onActivityResult. During that
time, input sends focus=false event to Activity B. But ActivityA is
occupying the main thread for the heavy computation, and therefore
Activity B is not able to respond to the focus loss event in time.

As a result, ANR occurs.

To fix this, run the heavy computation in a separate thread.

Here:
Activity A == GenerateImagesActivity
Activity B == ThemeDeviceActivity
heavy computation = compressOutput

Bug: 143459140
Test: atest ThemeHostTest --generate-baseline 4
Change-Id: Ib4c2bd6e71ef3266319ec7bbad2afae6bb29c083