Fix DreamManager test
The home activity should be invisible after the dream is started. The test was
asserting that this happens immediately, but in reality there is an
animation when the dream is starting, during which the home activity is
still visible. This CL changes all tests to wait for valid state and assert
visibility after that.

Bug: 156029327

Test: atest CtsWindowManagerDeviceTestCases:android.server.wm.DreamManagerServiceTests
Change-Id: I23fb03f1518bbb9b3817002dabd86ec695919dd0