Fix cts fail for PartitionedActivityTest
In the test we trigger the manual autofill request programmatically,
because the focus isn't on the field, we will not get the callback
from IME.
It would be better not to ask IME for inline request if the request
is manual and the view is not focused by appending a new flag.
For manual mode, it would be better to check if the flag of FillRequest
includes "FLAG_MANUAL_REQUEST" instead of checking equal.

Bug: 154661868
Test: atest CtsAutoFillServiceTestCases
Change-Id: I9b0151d41901f78cb8cc9ca5fa1db706d6d6dbe0