Remove check for prox sensor feature in batching CTS-V test
There is a logical error on the check, preventing the test to run
even if the device supports the prox sensor. Since the test will
"skip" the test if a sensor is not exposed by the sensors HAL,
remove the check to ensure the test runs if a valid prox sensor
exists.

Bug: 134477295
Test: Run test and verify pass
Change-Id: I2b91ef8031647145d26b76ba73eed8b8c28ad85c