Don't use metalava in Signature (device side) CTS
Using metalava in the device side is preventing from using "desktop"
Java features in Metalava. There was an idea of extracting metalava's
signature file parser into a library, but the code was so intertwined
that it wasn't easily doable, so we're going back to using the XML
API file in this test.

Bug: 151160048
Bug: 151224432
Test: ./runSignatureTests.sh
Test: Randomly change frameworks/base/api/current.txt and ...
 ... make sure `atest CtsCurrentApiSignatureTestCases` catches it.
Change-Id: I13d058895b3ec0e153008b96ad3fb3b0c0aa47c3