Resume-on-Reboot: make locking screen more robust
The test would try to send the SLEEP key, but that sometimes fails.
Additionally call the GLOBAL_ACTION_LOCK_SCREEN on accessibility service
to try to make sure the screen stays locked.

Additionally try to monitor the lock screen state and check that it is
visible after the lock screen is summoned.

Bug: 152359655
Test: atest ResumeOnRebootHostTest#resumeOnReboot_SingleUser_Success
Change-Id: I2887d7a15a62a5a3b8b86a7aadf158d1d0a11337