Fix compat changes for PackageSetInstallerTests
For -user devices, the packages need to be marked debuggable
so that they can be overridden.

The target SDK was lowered since a recent change made it so
that disabling changes are not allowed, only enabling.

The am compat call was moved after the package installation
since it doesn't support not-yet-installed packages.

The test names were also accidentally flipped in the params.

Bug: 158278000

Test: atest android.appsecurity.cts.PackageSetInstallerTest

Change-Id: I92f432e96b0da705920b1bb42ce41ae39f3cc1c1