[Signature CTS] Make sure to test enum fields
The jdiff output from metalava was by default in the "compat" mode
and didn't emit enum fields.

Disable the compat mode and make sure to test enum fields too.

The jdiff output didn't contain a signal to differentiate enum fields,
so also update metalava to add a marker for it.

Bug: 151224432
Test: ./runSignatureTests.sh

Change-Id: Ib02ecef05f2804b070e4d8de39d64b113a3011a5