Fix test breakage
Ibe113d4a4fe6fe51766e143513f76595cf43e594 introduced a change that
invocations to all methods on mIContentProvider throws an exception.

This caused an error when some of the methods are again mocked via

when(mIContentProvider.foo(...)).doReturn(...);

This is because "mIContentProvider.foo(...)" itself triggers the
aforementioned mock and throws an exception. The statement above
actually works since the second call to the foo method.

To fix the problem, above is changed to the following:

doReturn(...).when(mIContentProvider).foo(...);

, which doesn't call the method.

Bug: 159061894
Test: atest CtsContentTestCases
Change-Id: Icdb60510bfac39527d7cc993bb129e64614bdc41