Attempt to reduce test flakiness.
We've been seeing quite a few examples of flaky tests in this suite,
and one hunch is that previous test runs are leaving artifacts that
cause future runs to fail.  This change adjusts each suite to use a
unique "nonce" value in file and directory names, borrowing from the
strategy used in other CTS tests.

Bug: 159125970
Test: atest CtsScopedStorageHostTest
Change-Id: I0984bddd7f71677e7033d7e9075ccead6eb76359