STS test for Android Security CVE-2019-2092
This test is unique in that we are running under a new managed profile.
We try to call DevicePolicyManager.isSeparateProfileChallengeAllowed()
with USER_SYSTEM like an attacker would do. Since this test does not
run as USER_SYSTEM itself but instead runs in a managed profile, it
should fail with a SecurityException.

Test: cts-tradefed run cts -m CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.SeparateProfileChallengeTest#testSeparateProfileChallengePermissions
Bug: 130574934
Bug: 128599668
Change-Id: Ibde9cd8c49363a498699f68282013c58fc8a0cc0