Fix flaky InlineAugmentedAuthTest#testDatasetAuth_resultCancel
The test failed because we had an unhandled fill request. When the
Activity onResume() called, it may call notifyViewEntered() then
onFillRequest() will be called. The test pass when onResume() is
called after our check, otherwise fail. This is a timing issue, we
can find the view in the Activity to make sure the Activity is
resumed, and then pull the FillRequest.

Bug: 162426592
Test: atest --iterations 10 android.autofillservice.cts.inline.\
InlineAugmentedAuthTest#testDatasetAuth_resultCancel

Change-Id: If80b038c8c04d6eb2dd78a783c050296497df046