Make MotionEventTest work with a free-form window.
The test tries to inject a touch event inside the activity, but outside
of the overlays. However, it fails when the main activity's view is not
behind the overlay because a touch event is sent to the center of the
left edge of the left overlay.
This CL fixes it by sending a touch event to the center of the main
activity's view.

Bug: 162475453
Test: atest MotionEventTest on eve and on flame
Change-Id: I72b55e5d28675a286cb0366484c6fa0df6cdb64d