Revoke uri permission after moving document
There is an security issue. If the app got the uri permission before,
when the document was moved to other place, the uri permission was
not revoked. If the user adds new file with the same name on the same
location, the app will have the uri permission for the new document.
We should revoke the uri permission after the document was moved.

Bug: 157474195
Test: atest DocumentsTest#testAfterMoveDocumentInStorage_revokeUriPermission
Change-Id: I488f667cab80646ff2043a26f461d07a4ea7ea23