Deflake BackgroundCallAudioTest ConnectionServiceTest
BackgroundCallAudioTest#testLowerApiLevelCompatibility{1,2} was flaky
because tests that run before it can leave the CtsApi29InCallservice in
a bad state by sending a call to it and not resetting it via the control
interface afterwards.  Fix this flakiness by resetting the test
incallservice whenever Telecom unbinds from it.

ConnectionServiceTest#testConnectionServiceFocusGainedWithDifferentConnectionService
was flaky because the call to setActive, while delayed in Telecom due to
internal handlers, may get interleaved between the calls to
setDisconnected and removeConnection from the CTS cleanup. Fix this
flakiness by waiting on Telecom internal handlers before proceeding with
test cleanup.

Bug: 162647577
Test: atest, rerun until failure
Change-Id: Ie3ca7216930ffbdb80105e15e4434aa4216b3246