Fixes a flake due to unbind before the test could check
Sometimes, NonUiInCallServiceTest#testNullBinding would check to see
if the InCallService was bound after Telecom had already unbound due
to detecting that the object returned a null binding. This change
separates placing the call from verifying the call was added in order
to also verify that the InCallService was bound.

Also fixes an issue where when bind happened before the test checked
and the future was used, unbind would cause a crash.

Bug: 166490982
Test: atest --rerun-until-failure 20 CtsTelecomTestCases:NonUiInCallServiceTest
Change-Id: Icb44870abe11f9eab7050a4f9e4f1f402419ed87