Run previous action when failed to read ro.vendor.build.version.sdk
isVendorApiLevelNewerThan and isVendorApiLevelAtLeast
in PropertyUtil.java were changed to read property
ro.vendor.build.version.sdk and throw exception when
failed to read it. But it can be problem when run test
on old system image that can't read the property.
In that case, just return the result of previous logic.

Bug: 166800127
Test: gts-tradefed run gts -m GtsYouTubeTestCases
Change-Id: I633a8d2c82326d2f2e6a873ebfe29f39d1c10ef8