Go to keyguard after setting lock credential
The test is often failing due to the keyguard
not being shown after setting credentials. Since
the purpose of the test is to make sure the
keyguard is no longer disabled after setting the
credentials, go to the keyguard afterwards. If the
device is still disabled, the keyguard won't be
shown and the test will still fail as expected

b/168051810

Test: run cts -m CtsWindowManagerDeviceTestCases -t
android.server.wm.KeyguardLockedTests#
testDisableKeyguard_thenSettingCredential_reenablesKeyguard_b119322269

Change-Id: I2a44c64af3087eb9ed5fa06ca13dc1e5cf1201ff