Fix failing testForceDark_rendersDark test
The test is about how force-dark changes the background color of
WebView. It shouldn't involve the focus highlight at the beginning.
Previously the test could pass because the default focus highlight
was semi-transparent. But recently we changed the default focus
highlight to blue to support rotary, so the test failed. To fix this we
can make the WebView non-focusable so that the focus highlight won't
draw.

Fixes: 163814465
Test: atest android.webkit.cts.WebSettingsTest
Change-Id: I3d55b9727a808bcd00cc343386df8c06008ccd0e