Launch activities in fullscreen in IntentTests
This test verifies activity lifecycles, and the expected behavior is
that of fullscreen windowing mode. When the device default windowing
mode is freeform, the behavior is different. For example, in fullscreen
windowing mode, when one activity is resumed, other activities are
stopped, but in freefrom windowing mode, they are not.

This change uses ActivityOption to start an activity in fullscreen
windowing mode so that default freeform window devices pass this
test.

Bug: 168780834
Test: atest android.server.wm.intent.IntentTests on eve and crosshatch
Change-Id: I5bdceb2bb58d37177f30be32b253f3257a4fb534
(cherry picked from commit eeedf7b98eeb27d1e7e6aab5fc3fcc69f14e0e57)