Adjust split-screen tests to work on non-fullscreen-display devices
The CTS tests should pass on devices which have default freeform
displays (eg. ChromeOS). To do so, remove anything that specifically
relies on fullscreen in favor of "not-split-screen".

Also removed a test which can't pass on devices with close-to-square
aspect-ratios. The test would try to adjust the split divider and
look for an app relaunch -- however, this only happens if the app
actually changes orientation. On close-to-square aspects (like 4:3
pixelbook), it'd require making one of the splits unrealistically
thin and could even, itself, violate minimum-size constraints.

Fortunately, with the advent of task-organizer and multiwindow
windowing mode, this particular behavior should already be tested
by other tests

Bug: 168857697
Test: atest SplitScreenTests
Change-Id: I0ccf525e833e5425bfd325b57fa8c849304f3b5e