Reduce flakiness of testActivityStoppedWhileNextActivityNotIdle
If the testing device is slow, the main window of the non-idle
activity may not be drawn yet. But the transition animation has
done by starting window, and the resumed activity doesn't set
nowVisible to true. That causes the previous activity won't be
notified when the transition animation is done.

This change uses the test activities in the same task, so there
won't be a starting window to hit the timing.

Bug: 169508204
Test: ActivityVisibilityTests#testActivityStoppedWhileNextActivityNotIdle
Change-Id: I9febc50fb18e6f6d669c763b6738521ffb6858f9