AdaptivePlaybackTest: add timestamp to mRenderedTimeStamps before releasing buffers
After calling mCodec.releaseOutputBuffer(), onFrameRendered() callback
will be called when the buffer is rendered. Inside onFrameRendered(),
we try to remove the timestamp from mRenderedTimeStamps. However,
currently we add the timestamp "after" calling
mCodec.releaseOutputBuffer(). So the timestamp might be added after
we try to remove the timestamp.

This CL moves the timestamp adding prior to calling
mCodec.releaseOutputBuffer() to avoid the situation happens.

Bug: 171364305
Test: run android.media.cts.AdaptivePlaybackTest#testH264_adaptiveDrc
      and not seeing warning about timestamp

Change-Id: Ifc942351aae1a486ae4042ddf726766f41599e49