fix Music directory missing which cause gsi test fail
This bug was caused by Music directory was gone before this test. We
must confirm it exists first. If the well known root directory doesn't
exist,  we can create it in the beginning of the test. If Music and
Download directory is missing before the test, it will create them and
let test continue run.

Bug:153815767
Test: run cts-on-gsi -m CtsDownloadManagerApi28 -t android.app.cts.DownloadManagerApi28Test#testAddCompletedDownload_mediaStoreEntry

Change-Id: I94f49a0ea0321a171ca69f3cca6c76ba22489089