Refactor HideApplicationOverlaysTest
Put the test activitiy into a seperate proccess. This is done to
prevent the reuse of existing Session object in the com.server.wm.cts
process. Session objects cache the permission state at creation time,
this is fine for "real" usage, but since this test is adopting the shell
permissions it changes the permission state without restarting the
process. This isn't possilbe during normal use. To work around this, we
ensure that the test activity is always run with a fresh pid, to
guarantee that Session is created with the permissions the test wants.

For example if another test was run first that didn't have
INTERNAL_SYSTEM_WINDOW then the Session would "remember" this, and
testInternalSystemApplicationOverlaysNotHidden would fail because the
permission state of Session isn't refreshed when the shell permission is
adopted.

Note: Adopting shell permission works per uid, which is why we keep
our helper activity in the main test uid, but in a separate process.

Bug: 175816428
Test: atest CtsWindowManagerDeviceTestCases:HideApplicationOverlaysTest
Change-Id: Ib6fb81ae5745361d87b3624aa40c3999efff5e80