CtsNdkBinderTestCases: test transaction binder
Adding test case to support making new changes - to ensure this behavior
isn't broken. AIBinder_transact rejects transactions that are made, when
it is passed a parcel which was prepared with another binder (this
behavior is also documented in binder/android_ibinder.h). While this was
defensive before, it turns out it is a very useful restriction when
binder may be used with different wire protocols.

Bug: 167966510
Test: atest CtsNdkBinderTestCases
Change-Id: Iea856ce43ab96a760100491bed4b236a48809aa1