Add CTS for animation touch exemption
For entry animations we use ActivityOptions for launching the
OverlayActivity with a local animation to the test. For the exit
animation we create a new activity in the helper app that doesn't have
windowIsTranslucent equals to true, since the exit animation of those
doesn't seem to honor custom-defined animations. Then, in that new
activity we use overridePendingTransition() when we get the finish()
signal.

Refactored a bit to be able to inject input while animating.

Bug: 158002302
Bug: 172787052
Test: atest WindowUntrustedTouchTest
Change-Id: I17ce2f237a81dc08a6cf5d2df86f30c208221a8b