Move InstallUtilsHost to the android package
This is a partial cherry-pick of the change
I141441801623417d053452b500a1e064e7304184

Packaging the library under android.* package helps avoid
that. This is a temporary measure until we figure out what is the
correct way of adding a host utility library.

Test: atest sdkextensions_e2e_tests
Change-Id: I1afae43bc29bd3d1ff254593c7ceb12f4b59b25a
Merged-In: I141441801623417d053452b500a1e064e7304184