Test for NPE in PackageProperty class
Reported by zezeozue@; if no package defines an application
<property> tag, getProperty() was throwing an NPE within the
package manager service.

NOTE: There's now a CTS test to ensure this works. However, it's
not sufficient. As soon as there's a single system application
that defines an application <property> tag, the test will
no longer catch the condition. We need to re-write the CTS
test as a Package Manager Service unit test and verity the
behaviour outside of the running service.

Bug: 169258655
Test: atest CtsContentTestCases:PackageManagerGetPropertyTest#testPackageRemoval
Change-Id: Ic29aa916727d489c2e687a011d36a1747cec687e