ScopedStorage: Error message if startsWith fails
When testing the FUSE passthrough patch set I noticed that
CtsScopedStorageDeviceOnlyTest was failing at
testOpenContentResolverFirstWriteFilePath, but no error message was
printed to help figuring out what was the actual value and what
expected.

Changing the startsWith assertions to print an error message in case of
failure that tells the evaluated string and the expected prefix.

A similar change for CtsScopedStorageCoreHostTest has been recently
merged.

Bug: 168023149
Test: CtsScopedStorageCoreHostTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I8a2449bd56b272c07373d9dd72ac74e7a1ed8aaa