Refactor WindowUntrustedTouchTest to reuse code
Turns out module cts-wm-overlay-base is a dependency of the tests. I
wanted to do this in this CL but I ended up doing in ag/13164095 and
didn't realize (since that was a prior step to this one). Anyway, this
means we can reuse everything from cts-wm-overlay-base instead of
creating our own SAWs, activities, etc inside the test.

Also, created a new AIDL interface for the bound service that provides
interaction between test and test app. This removes duplicate code,
removes boilerplate code to pack/unpack messages, has a nice type-safe
API to use and benefit from the synchronicity as well. Also was able to
reuse service methods inside the test for the test app itself.

Created a new shared module between the test and the base test app to
contain the AIDL which can be reused for other shared stuff between test
and test app.

This was done not only because I think it's a good clean up but in
preparation to add a test for b/173317413 in which I'd have to recreate
an activity inside the test.

Bug: 173317413
Bug: 158002302
Test: atest android.server.wm.WindowUntrustedTouchTest
Change-Id: Ibdd972a4e683feb1b4a50caed672f36eda47d1e2