Fix/refactor WindowInputTests
* Most of them were making assertions inside the listener but were never
verifying that the listener was actually called.
* In testFilterTouchesWhenObscuredByWindowFromDifferentUid() the
listener was not meant to be called at all, so removed it.
* Created another testFlagTouchesWhenObscuredByWindowFromDifferentUid()
that verifies touches are flagged correctly.
* Use the maximum allowed SAW opacity for go/untrusted-touches where
necessary since we are verifying filter touches API here, not
go/untrusted-touches. The effect is the same, that's why the tests
didn't break with ag/13228932

Bug: 158002302
Test: atest WindowInputTests
Change-Id: I7cd3c3494e6b1414d5a78278c5181717efb3a6df