Fix flakiness in AppPredictionServiceTest due to async.
Calling AppPredictor#registerPredictionUpdates immidiately followed by
AppPredictor#requestPredictionUpdate could sometimes leads to a
timing issue where the former hasn't been running to a complete.

In production the later is called in the callback of the former, which
already ensures sequential execution. However since the same is not true
for this test case, we need an extra delay in between to mitigate test
flakiness due to the timing.

Bug: 166669976
Test: atest CtsAppPredictionServiceTestCases:\
android.apppredictionservice.cts.AppPredictionServiceTest#\
testRegisterPredictionUpdatesLifecycle --iterations 10 -- --abi arm64-v8a

Change-Id: If19df9473b6eb37707ecb50758180789e28360f5