Allow hidden APIs during CloseSystemDialogsTest
We need to test that a few @hide APIs are properly protected. The
default helper app (android.app.stubs) does have the ART checks
disabled. However, we don't use that app, we use com.android.app1 on
purpose because android.app.stubs has special powers due to the fact of
being instrumented and we're verifying the absence of those powers in
our test.

So, disabling the check globally for the duration of our test.

Bug: 177359344
Bug: 159105552
Test: atest CtsAppTestCases:android.app.cts.CloseSystemDialogsTest
Change-Id: I3afff94691bf64b92958909c903a1f54b35d7993