Adding tests for device ID attestation
These tests check that the getters and setters for device ID
attestation work as expected, now that generating and attesting keys
occurs in the same call.

A few other fields that had been added to KeyGenParameterSpec but not
simultaneously added to the CTS tests have also been included in this
test for correctness.

Test: atest cts/tests/tests/keystore/src/android/keystore/cts/KeyGenParameterSpecTest.java
Bug: 177369988
Change-Id: I805fc296cecbbfdd4c99195787f632c60496f526