VoiceInteraction tests: Handle null args in onshow().
Due to a concurrency bug in VoiceInteractionManagerService, the args may
be null even when the test code sends a non-null value in the request.
This was causing the service to sometimes crash, which was causing the
tests to fail with timeout errors.

The arg can also normally be null if an app doesn't provide a Bundle
when calling Activity.startLocalVoiceInteraction() or
Activity.showAssist(), so it's good to handle this in the test service
anyway.

Bug: 178325554
Bug: 178776303
Bug: 178775773
Test: atest CtsVoiceInteractionTestCases --iterations
Change-Id: Iff35dbae67a1e4cd4751230d47aa5b63e7d95443
(cherry picked from commit f42245a33084f34a265acdf8d7d507f7afca2e3a)