Add test for rename File API support for media URI grants
Add test for rename File API support for media URI grants by
createWriteRequest. A file for which the caller has write-access due to
the virtue of createWriteRequest can be renamed to any file (provided it
follows scoped storage restrictions and the new file does not already
exist). But the caller may lose access on the new file as the media-URI
changes. Please see b/178816495 for more details.

Bug: 173648980
Test: atest ExternalStorageHostTest
Change-Id: I1987a8697eced0933f66834cc97ecada86f7955b