Camera: Wait for capture callbacks
Capture process related callbacks such as 'onCaptureProcessStarted'
and 'onCaptureSequenceCompleted' are not guaranteed to arrive
before or immediately after the extension processed image. The
callbacks are always triggered after the respective 'process'
method is called but they will run in a separate thread.
Depending on timing such as extension processing latency and
scheduling, the processed image could arrive well before the
callbacks are able to reach the client side. Account for this by
waiting on the respective callbacks with appropriate timeout.

Bug: 179450326
Test: Camera CTS
Change-Id: I09f72cccc6144719304f0304504421a29dfad9fa