Start of DeviceAdminReceiver support in EventLib.
This does not include full tests, because they will require manipulating
device admin. These will be added after the Nene support for device
policy is merged (CL pending).

This will also be followed by a refactor so that Activity,
DeviceAdminReceiver, and BroadcastReceiver CONTAIN a class rather than
BEING a class. This will make it more scalable as we can add
class-specific filters without having to implement them in multiple
places.

Test: atest EventLibTest
Bug: 180604880
Change-Id: Ib5a0889fa59f1f6cb40e34a91857bbb5784d31d9