Turn off instant mode for GlobalSearchSessionPlatformCtsTest
GlobalSearchSessionPlatformCtsTest is setup to sideload two helper apps
onto the device, then bind to their service to trigger the apps to index
some documents. This lets us test with multiple packages. Unfortunately,
there are restrictions on how packages can communicate when in instant
mode, which causes this test to fail in instant mode. Since the setup is
a bit special and the functionality is still being tested in non-instant
mode, turn off instant mode for this test class.

Bug: 180407077
Test: atest --instant CtsAppSearchTestCases
Change-Id: I690ca405bc9b7a16b245925e85cb57d86411324d