Mark accel/gyro/mag range tests with AppModeFull.
This is needed because sensor cts package has the
HIGH_SAMPLING_RATE_SENSORS permission. However, this permission is not
in the set of permissions that instant apps can have. As a result, if
the sensor cts package is installed as instant apps, it would appear as
if it does not have the permission, hence failing these range tests.

Test: cts-tradefed run cts -m CtsSensorTestCases -t android.hardware.cts.SensorParameterRangeTest
Bug: 136069189
Bug: 181186218
Change-Id: I17312644fbaf28e8578494c768082901b97c0b93