Fix Error With testOnTouchEventHorizontalMotion
This test is to verify whether the scroll distance
is smaller than rightMost.Scroll distance is equal to the
min(config_viewConfigurationTouchSlop and rightMost).
Since we changed config_viewConfigurationTouchSlop from 8
to 48, and rightMost is based on the length of content in
TextView (in current case it is "hello world"). Thus this case fail.
Solution is extending the content according to
config_viewConfigurationTouchSlop.

Test:run cts CtsTextTestCases
bug:171755304

[DO NOT MERGE]

Change-Id: I51e1dc5c8efb35c27c9fbe7b306eb1ae0f075f9b