Highlight difference between isAtLeastS() and isPlatformOrWifiModuleAtLeastS()
In CTS, BuildCompat.isAtLeastS() should be used
to gate an API when it has dependencies on
S HALs/APIs. It never makes sense to call such
an API on a R platform even if it has the S
module. If it were called, it would throw an
UnsupportedOperationException.

WifiBuildCompat.isPlatformOrWifiModuleAtLeastS()
should be used to gate an API where it does not
have any dependencies on S HALs/APIs. Such APIs
are callable on Android S, as well as Android R
devices with the S module installed.

Bug: 181358771
Test: compiles
Change-Id: I8b1728a1ffdb7dbdccf1d6acab7776adc634bd58