StrictJavaPackages: add exemption for WearOS duplicates
Some of the duplicate classes in system server class path come from
wear-service.jar. These are not easy to get rid of since they are
statically linked from vendor libraries. The amount of these classes is
rather small so the suggested workaround to pass CtsStrictJavaTestCases
is to add them to the exemption list.

Bug: 184018499
Test: atest -c android.compat.sjp.cts.StrictJavaPackagesTest
Change-Id: I71c28371f2675093247a0be9c95fefeea7aa979f