Fix secondary user UninstallExistingPackageTest
The test testUninstallExistingPackage_failsIfInstalledInOnlyOneUser
tests that packages that are only installed in a single user cannot be
completely uninstalled with the uninstallExistingPackage API.
If the test was executed with `--user-type secondary_user`, tradefed
created a new user to run the test in, and the target_preparer installed
a test APK in all users (primary and the created secondary user).
Thus when uninstallExistingPackage is called, the package can be
uninstalled. The test did not expect this case.

With this change, the test apk is only installed for the user the test
is running for in setUp, and uninstallExistingPackage is also called on
that user instead of always the primary user.

Bug: 180755384
Test: atest CtsPackageManagerMultiUserHostTestCases
Test: atest --user-type secondary_user CtsPackageManagerMultiUserHostTestCases
Change-Id: Ie1963cbb8d3c4a1c0ed012482c578d3043766b99