instant apps don't have access to the KeyguardManager
Call into KeyguardManager, or set FLAG_DISMISS_KEYGUARD
Call setTurnScreenOn in conjunction with showWhenLocked, to make
sure the activity is visible when the lockscreen is up.

If the screen is on, FLAG_KEEP_SCREEN_ON will ensure the screeni
stays on and is bright as long as the window is visible.

Bug: 185788282
Test: run cts -m CtsAccessibilityTestCases -t
android.view.accessibility.cts.AccessibilityEventTest

atest --instant AccessibilityEventTest
Change-Id: Iac8ef7bc10f0ea64b33aff3607beddbd70164321

Change-Id: I979650f1daf2160d9aaedfdcb2db51882e3fd474