Fix lint warnings under cts/tests/inputmethod/ (again)
This is a mechanical code clean-up, which follows up to our previous
CL [1].

Although we have enabled lint-check during "repo upload" [2], it
unfortunately doesn't work for CLs that are cherry-picked on gerrit.
We still need to pay attention to lint errors' creeping into the code,
especially when CLs are cherry-picked on gerrit.

There should be no change regarding what's asserted in these IME CTS
tests.

 [1]: I57c24911391c6454b5f7414be238839089256053
      65174874b58f64bb8db50dc39a4c48fd2f2b8c13
 [2]: Icafcdcf653315ef4b67df88bba31ca2f5af2de59
      1342dbc7057d897ac485d5b27eb5fd6cb377e61f

Bug: 171827834
Bug: 173079401
Test: prebuilts/checkstyle/checkstyle.py \
          -f cts/tests/inputmethod | grep .java
Change-Id: I0b8603feba8c1e8b6e371e821c7b5f983325088b