Fix exception handling in WebViewSyncLoader.
If WebViewSyncLoader.pumpMessages() exits early due to a task throwing
an exception that is *not* the specific ExitLoopException we are waiting
for, the task that throws the ExitLoopException will remain on the queue
and can crash the actual non-nested event loop if it gets to run before
the exception that was thrown crashes the test anyway (which can happen
with @UiThreadTest as the test thread and UI thread are racing in that
case).

To avoid this, remove the Runnable we posted from the queue before
exiting the function, to ensure it can't possibly run after we return.

Fixes: 188588550
Test: atest WebViewHostSideStartupTest
Change-Id: Ifccc146a7517d06cc65934a23ca1ae2255f87ef1