Use waitFindObject() instead of waitFindNode().
waitFindNode() doesn't support scrolling and causes the test to fail on
shorter screens. Clicking the "Permissions" preference in app info page
is commonly done with waitFindObject() (even in other places of this
test) so there shouldn't be any problem. We can't remove waitFindNode()
for now because it is used elsewhere.

Fixes: 187879488
Bug: 188567164
Test: AutoRevokeTest
Change-Id: I30e09d2b991c5c189e7a1a4c323f645bbffcd4f1