Fix Cipher Test is failing between Peeking display and Keyguard windows
Reason: CTS uses the DUMPSYS WINDOW command to get the current state of
the windows. Even the screen is off, the AoD and Keyguard flag controls
returned by that command are true, so the keyguard state is verified
erroneously when the AoD is still visible. This causes the CTS to send the PIN
while the device is showing the AoD screen.
 
Fix: The condition to assert the keyguard is visible was changed to also
verify the AoD state. Now, for CTS purposes, the keyguard is only
accepted as visible when its flag is true and the AoD flag is false.

Test: Attest android.keystore.cts.CipherTest

Bug: b/189420672
Change-Id: I6df97683a8f4c30959dfdd3d4176398770ada240