CtsMediaV2TestCases: Fix codec-prefix support
Recent changes when parameterizing the tests further,
codec-prefix support was broken.

Instead of selecting codecs starting with the given prefix,
those were getting rejected.
Also codec create was attempted before selecting/rejecting
the codec, which is unnecessary.

This is now simplified and codec-prefix is used to select
codecs to be tested.

Bug: 192107461
Test: atest CtsMediaV2TestCases -- --module-arg \
 CtsMediaV2TestCases:instrumentation-arg:codec-prefix:=c2.android.

Change-Id: Iaf7ff3ed45392211e23bf0a1dd645d6c9876a103