Avoid calling verifyTest multiple times in ASurfaceControlTest
Every call to verifyTest will re-create a new SurfaceView, create a new
SurfaceControl, and then tear it down. This can be very slow and create
issues with running the test. Use MultiRectChecker to verify multiple
colors in a single SurfaceControl, rather than calling verifyTest over
and over. This seems to remove some of the flakiness from these tests

Bug: 191613127
Test: ASurfaceControlTest --rerun-until-failure
Change-Id: I03c26478c12baea15a0bbe90d8d7a9b2a0154e83