cts_aaudio: skip basicMMapTest
The test tried to verify that an MMAP stream could be opened on a
device that claimed to support MMAP.
But the test had problems.

It was failing for some partners for various reasons.
We can enables it in T if we add a HAL query for MMAP support.
We also need a test function to see if MMAP is supported for a specific
device.

Bug: 192313064
Bug: 192984235
Test: Run AAudio CTS test suite using tradefed.
Test: It should skip test_aaudio_mmap_testBasicMmapOutput
Change-Id: I1322355abc3f646c6743e3bb09f7d5e43469e9c8