Fix LockTaskTest.
LockTaskTest was defined incorrectly and was only passing due to a
separate bug which meant it (usually) doesn't test what it should!

This redefines LockTask as only applying user-wide (which appears to be
correct) and fixes the bug which was causing it to incorrectly pass
before (it wasn't setting the primary dpc correctly).

Test: atest CtsDevicePolicyTestCases
Fixes: 193513706
Fixes: 194092066
Change-Id: Icea0f21ba3802b69479c2994fd4346182963be48