Fix WindowLeaked warnings in logcat from ImeInsetsVisibilityTest
This is similar to the fix I recently made for FocusHandlingTest [1].

If you manually create a window by using WindowManager#addView(), then
you are responsible for destroying it with WindowManager#removeView()
before finishing the application.

This CL enforces it by using try-with-resources statement.

Other than that, there should be no semantic change in
ImeInsetsVisibilityTest.

 [1]: I11c193bc78627cc46d3409d1ae85aec1eef92c2b
      34a248dca96835f1de3518e0eb1beeaa6d441103

Fix: 193826172
Test: Manually verified as follows
 1. adb logcat -c
 2. atest -c CtsInputMethodTestCases
 3. adb logcat -s WindowManager:E
    Make sure that there is no WindowLeaked warning
Change-Id: I78beea01f616c5241551ee0bf67c0eec9a0310aa