Increase timeout again for running location access check job.
It is observed on Coral devices that such a job may take around 5
seconds to complete, which is the current timeout, and it is causing
flakiness. For instance, in one bugreport a location access check job
was started at 11:16:07.214 and only finished at 11:16:12.676.

Bug: 182960931
Test: presubmit
Change-Id: I6beaf163502f3fe421b02f37f14062332bfcaf64
(cherry picked from commit 3933ed580ab03a66ba92d14db041a705b2bdc428)