Fix WindowInsetsAnimationTests for PC CF
To hide and show systemBars in WindowInsetsAnimationTests, the test
activity must be in fullscreen. In freeform windowing mode devices
(like PC CF) the activity doesn’t cover the systemBars so all
WindowInsetsAnimationTests are failing because of that. The solution
is to run the activity in fullscreen windowing mode. But this
introduces a new issue and that is the height of the captionBar that
is not zero in freeform mode.
CaptionBar for fullscreen devices always has zero height and because
of that it doesn’t have any effect on systemBars. But for PC CF, the
captionBar is not zero and that breaks WindowInsetsAnimationTests
assertions. To fix the issue the captionBar must be excluded from
systemBars. This will not have any effect on fullscreen devices.

Bug: 195339488
Test: atest WindowInsetsAnimationTests
Change-Id: I15b2a5f83ced6ec6fce19e396a4c5b7d11016a01