Make BlurTests use px instead of dp
This CL removes the usage  of dp from the BlurTests.
This is because the Kawase blur algorithm is only an approximation of the
Gaussian blur and it does not grow linearly. So we can't make reliable
assumptions for what the blur effect will be on devices with different display
densities.
Converting to pixels makes the test perform over the same amount of
pixels on all devices, so the result (in pixels) is always the same

Bug: 193861404
Test: atest BlurTests
Test: run the tests with different densities: `adb shell wm density 420`
Change-Id: I6feea47c467e024a3e74686a0ea4a85c2a5e168e