Fix flaky InlineSimpleSaveActivityTest#testAutofill_oneDatasetAndSave
In the test, we will change the EditText text and then call commit().
The test fails because we call commit() too early. Because the value
doesn't change yet, the saveui will not show.

To fix this issue, the test will wait the EditText value change then
call commit(). Without the fix, there are 40% to see this issue (not
fail test but pass on re-try). Local verify solution in 100 tries,
not see the problem again.

Bug: 195582566
Test: atest InlineSimpleSaveActivityTest --iterations 100

Change-Id: I6970e4114a1efacc36e25871b77c5350496a0b2a