Fix UiAutomation tests in 720x1280 resolution
This fixes several UiAutomation tests by doing the following:

1. Adjust overlay handling in PermissionTapjackingTest
2. Increase timeout for testNoResidualPermissionsOnUninstall
3. Scroll past CollapsibleToolbarLayout when detected
4. Fix end-of-scrollable-view detection

Points #1 and #2 are test-specific issues.

Point #3 addresses an issue that affects many, if not most,
permissions-related UiAutomation CTS tests.

Recently, CollapsingToolbarLayout was introduced, which broke most of
these tests when they are run in 720x1280 resolution. The root cause is
that many of our tests currently rely on
android.support.test.uiautomator.UiScrollable, which wasn't designed to
handle CollapsingToolbarLayout.

The problem with CollapsingToolbarLayout, and the reason UiScrollable
can't handle it is that, if you try to scroll down, but you don't scroll
down far enough, the scrollable view just snap back to the top. This
means CollapsingToolbarLayout has a "swipe threshold": you have to swipe
your finger at least a certain distance (which seems to depend on length
of the heading) in order to gain any scrolling progress, or it will snap
back to the top.

So, the solution is: whenever a CollapsingToolbarLayout is detected on
the screen: calculate its height and scroll past it.

Point #4 fixes the following issue: scrollable.scrollForward() is
documented to return false when it has hit the end of the scrollable
view. However, it isn't working reliably. So, the workaround is to look
at the scrollable view's bounds before and after we call scrollForward,
and compare to determine whether we've reached the end.

Bug: 194146521
Bug: 192829458
Test: atest PermissionTest22 PermissionTest23 CtsPermission3TestCases
Change-Id: I9a29ab88d070db6161435f84b557665f8eb93539
Merged-In: I9a29ab88d070db6161435f84b557665f8eb93539