CTS tests for new API
These tests are for the new API for requesting that a TileService be
added to Quick Settings.

The automated CTS tests check failures that happen in the app process or
system server before the dialog is sent to the user.

The verifier tests check the user interaction part of the API.

Test: atest RequestTileServiceAddTest
Test: CtsVerifier Tile Service category
Fixes: 197207978
Change-Id: Ie953e89e200eea1cdeb97113697ae316d62be371