Add end-to-end tests for bad IC text-retrieval calls
This is a follow up CL to our previous CL [1], which added CTS tests
to ensure that negative lengths passed to surrounding text retrieval
InputConnection APIs do no longer cause process crashes in the IME client
side in a unit testing style.

With this CL, the above scenarios will also be tested in an end-to-end
manner.

 [1]: Ief79e62a7ca33c7cab6ea6dfb34c3dc4fc78bc84
      d361621dea21aab58169c0d7377fa70e0ec1728b

Bug: 169114026
Bug: 194110780
Test: atest CtsInputMethodTestCases:InputConnectionEndToEndTest
Change-Id: I89246d7847a4301d6396a1b17c24588b021888c7