webkit: Fix ProcessFreshChecker.
ProcessFreshChecker was never setting the fresh process flag to false,
so wasn't actually checking anything. Correctly set it to false after
the check so that the tests which depend on TestProcessClient will fail
as intended if the service process they create is not freshly launched.

This also revealed that PacProcessorTest is not properly closing the
client interface after each test and so the process was being reused
sometimes; fix this as well.

Add some docs to TestProcessClient explaining what it's for.

Fixes: 197796143
Test: atest WebViewDataDirTest PacProcessorTest
Change-Id: I7632b53f6888bd4c5e70cd68ca36e2617081d9f5