Ensure Collapsing Toolbar without scroll is collapsed
When an app has a collapsing toolbar, but not enough content to scroll
once the toolbar is collapsed, the UiAutomator will not know how to
scroll past it. If a view isn't found in waitFindView, and there is no
scrollable, check for a collapsing toolbar, and, if found, perform a
swipe from the bottom quarter of the screen to the top.

Test: atest PermissionTest23
Fixes: 192829458
Change-Id: I0e2ae1639f31f0dfc03e9ea0a6442973b24820eb
(cherry picked from commit dc3ec703c3f10420546e3f52fe9ff9d7c4651231)