Fix the test case "testHotwordDetectionService_concurrentCapture"
Due to the mDetectionJob Runnable doesn't be clear before testing this
test case, it will interrupt the test flow.

The root cause is :
We have a test case to test process died. After finishing that test
case, the ActivityManager will schedule a timer to restart
MainHotwordDetectionService. In following test cases, it may meet
that the ActivityManager restarts MainHotwordDetectionService again
during the testing.

It should reset mDetectionJob and mStopDetectionCalled when
MainHotwordDetectionService is initializing.

Bug: 201380107
Bug: 201531026
Test: atest CtsVoiceInteractionTestCases --iterations
Change-Id: I5ac80c99332a3334fce248d373b12d98dfe1c48b