Use Integer Property Animator for AttachedSurfaceControlTest
The current test uses a float animator to set the view's position.
However, the SC position is set with the setGeometry call which only
takes integer. This means the SC's position and the view's positions can
be off by floating points since the view's position will be converted to
an int when setting the SC its position.

Use an integer animator so we don't get floating point offsets

Test: AttachedSurfaceControlSyncTest
Fixes: 201764183
Change-Id: I4c9a6e5af4089223d7d2f669bf510be11f52f415