Use FIXED_SOURCE compatibility in MatchContentFrameRateTest
The match content frame rate feature is mean to be used video
surfaces, so we should use fixed source compatibility when
tesing. The difference is subtle, but in certain cases it
makes a difference. For example if the current refresh rate
is 59.94 and a default surface requests a change to 60 and
the device doesn't support seamless switching, the framework
won't trigger a mode switch because the cost of black screen
is higher than the benefit.
But if the surface uses fixed source compatibility the
switch will be made.

Fixes: 204300088
Test: atest MatchContentFrameRateTest
Change-Id: Icf73abca686e3e81f683b9f112d17b8d7d1ac5c3