Wait for PIP callbacks before asserting them
In some PIP-related tests, we assert that some callbacks are called
properly, but as they are called in an async fashion, we need to
wait for them to prevent timing issues.

In some other test cases, we already wait for such callbacks
properly, so probably we just forget to do so in these tests.

Bug: 206362738
Test: PinnedStackTests
Change-Id: I19162a0752a8cd3057f29634eb1a743bba471d90