Signature tests: catch and log NoClassDefFoundError
Avoids failing the tests if a NoClassDefFoundError is detected when
attempting to retrieve a class or its constructors, methods or fields
using reflection. Instead log the failure and continue.

This should not make the tests any less rigorous or risk allowing
access to implementation details of the platform as if the test cannot
retrieve them using reflection then an app will be unable to do so
either.

(cherry picked from 0c4d2d082b22dfc0ed6087de6bae9c08f43d057b)

Bug: 206650495
Test: m CtsSystemApiAnnotationTestCases
Change-Id: I94716a359aaa8c88429143130566ec4db13a7ad8