Fix CtsSettingsHostTestCases
In launchPrivacyAndCheckWorkPolicyInfo(), we launch
Settings.ACTION_PRIVACY_SETTINGS and then check if the work policy item
shows up in the screen.

But this implementation is not robust:
A false negative would happen if the work policy item exists
in the list but is hidden in the invisible part of the scrollable list
when the whole list does not fit into the screen.

To fix the problem, let's explicitly scroll to the work policy item.

Bug: 201245877
Test: atest com.google.android.cts.settings.PrivacyDeviceOwnerTest

Change-Id: I49a6f24e7e57f10c93c5709860b4f6f17bc209e1