Set activity windowing mode to fullscreen in WindowInsetsControllerTests
Many test cases in WindowInsetsControllerTests wait for status bar
and/or navigation bar bing hidden until time out. However, when
TestHideOnCreateActivity is launched in freeform, the
WindowInsetsController of the activity window can not hide the two
system insets if the window is not overllaped with the insets. Thus, the
test fails due timeout while waiting status bar to be hidden. The
activities of two test cases are not launching in the fullscreen:
- testWindowInsetsController_availableAfterAddView
- testHideOnCreate
Launch them in fullscreen to fix the potential issue.

Bug: 196731312
Test: atest WindowInsetsControllerTests on devices which set default
windowing mode to freeform

Change-Id: Ib690a1fa9491477c60335a2bcf57b50aa9937abc