Allow multi-window environments to place indicator anywhere
The current privacy indicator bounds tests requires that
the bounds be placed at the top of the screen since that
is the highest priority area in phones.
This CL allows OEMs to place privacy indicators anywhere
on the screen in multi-window environments, where the
highest visual priority area may not be the top edge (e.g.
desktops)

Bug: 213882561
Test: atest PrivacyIndicatorBoundsTests
Change-Id: I94ef3fb85a9e28d681e852419efb4a726f32b5d4