Add tests for the LocalSocket(FileDescriptor) constructor.
This CL also stops running the test for the legacy
LocalSocket.createConnectedLocalSocket method that the new
constructor replaces. This allows us to delete that method.

There is no impact to test coverage on previous Android releases
because LocalSocket is part of the platform only until S, and
does not need to be tested by T CTS.

Bug: 200200870
Test: atest FrameworksCoreTests:android.net.LocalSocketTest CtsNetTestCases:android.net.cts.LocalSocketTest on AOSP
Test: atest CtsNetTestCasesLatestSdk:android.net.cts.LocalSocketTest on S device
Change-Id: I0fd8a98088dedd82170d80d0080e375600595a13