Add coverage for setFrameRate(float, int) API
API 31 added a new setFrameRate(float, int, int) API
that wrapped the API 30 API setFrameRate(float, int) and with it
the cts was modified to use the new API 31 version. This caused the
API 30 version to loose API coverage. In this change the cts calls the
API 30 version if possible, and API 31 otherwise.

Test: atest SetFrameRateTest
Fixes: 214292297
Change-Id: Ia93fc224419d385946983c8b29aa29ef7b7162c3