[RESTRICT AUTOMERGE] CtsSystemApiSignatureTestCases: Stop checking android.test.* classes
A previous change has caused CtsSystemApiSignatureTestCases to have a
runtime dependency on android.test.mock which breaks the test as it
checks to make sure that those classes were not accessible. The reason
it did that was to ensure that those classes were not available by
default.

As that is also tested in CtsCurrentApiSignatureTestCases this removes
the check from CtsSystemApiSignatureTestCases.

Test: atest CtsSystemApiSignatureTestCases
Bug: 215797058
Change-Id: I057e6d8788d8d6fbffd5c5b62e2a0ffd9758bb0c