Add required flag to registerReceiver for T+
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
non-system broadcasts. This commit adds the RECEIVER_NOT_EXPORTED
flag to the receiver in HeartbeatActivity so that the corresponding CTS
test ActivityManagerAppExitInfoTest won't fail.

Bug: 218922295
Test: atest android.app.cts.ActivityManagerAppExitInfoTest
Change-Id: Ib2482fa4aef80d3fae185c452a68b800bc662c1a