add VoiceInteractionMultiDetectorTest CTS test
This test incorperates a new testing interface for CTS tests interacting
with a VoiceInteractionService. The new service adds a binder interface
for test classes to communicate with. The intention of this bindable
service is to provide proxy interfaces to the classes which are normally
only available to a VoiceInteractionService implementation.

Test: atest VoiceInteractionMultiDetectorTest
Bug: 193232191
Change-Id: I89eef0121ba5623d13e51c7cc9ed91a0a9fa56a2