Revert "Fix flaky AccessibilityMagnificationTest"
This reverts commit 06a392f18b81aa0925ffc04b3b2a08c7b1fa9a1e.

Reason for revert: In Android Auto, CarLauncher is the HOME activity
(com.android.car.carlauncher.CarLauncher). CarLauncher is based on
TaskView, therefore the active window will be maps (or any other top
most activity in TaskView) and not CarLauncher.

For #isHomeScreenShowing to work on auto we need to check if
CarLauncher's root window is visible to user. If it is then we have a
guarantee that home screen is showing.

Change-Id: I521179c2a1c6885c2c9c9973c5f9216ac5f12774