Solve the problem of the test item test_multi_camera_aligment
A previous problem with this test project was that circle in the
ultra-captured images were not recognized. The analysis found that
because the FOV of ultra is too large, the circle in the image are
too small, and it will be filtered out. (The script will filter out
contours with smaller areas)

The current solution is to relax the filtering conditions and make
the minarea smaller.

bug: 214881385
Signed-off-by: kongchenrui <kongchenrui@xiaomi.com>
Change-Id: Ic6012346ad256e611b588853845267f2dd0ca31b