Fix top and bottom edge extension tests
The test app was not an edge-to-edge app so the app window contained black bars at the top and bottom to account for the status bar and nav bar insets. But this meant top and bottom edge extension were black extensions which is not what the test was testing for.

Also made some changes to ensure the tests account for variations in color conversion better.

Bug: 224617933

Test: atest CtsWindowManagerDeviceTestCases:AnimationEdgeExtensionTests
Change-Id: I47ec5ec34aa8f9b6aaf3dc382c0eac17d8e1dca8