Make autofill CTS tests use sticky broadcasts
This CL uses sticky broadcasts to send the setInlineSuggestionsExtras
command to the MockIme in autofill CTS tests.
This is needed to handle the IME not running at the time of sending
the command, which can be the case when e.g. the
config_preventImeStartupUnlessTextEditor option is set.

Test: atest AutoFillServiceTestCase
Bug: 226240255
Change-Id: I42b1af2769cdd09d2c749f975b90c4710d2a48f6
(cherry picked from commit 090ee7d4bde9eb3072bf617ead740ddf1a8211bb)