Make 2021 CTS SIM required for T.
This SIM spec has been available since S, but we allowed a migration
period for one release cycle to ensure all testers now have updated
SIMs. The 2021 SIM is backwards compatible with the "legacy" CTS SIMs,
and includes its certificate as well as a newer one.

We update the CarrierApi suite's failure message to very clearly
indicate when a legacy SIM is still detected on T+ tests and point to
the 2021 spec.

Bug: 179534693
Test: make, use keytool to inspect test APK signature
Test: atest CtsCarrierApiTestCases (with new + old SIMs)
Change-Id: Ic61bc8178e5fb736b0652c22964d856ef46371ac