Separate SignatureTest and SignatureMultiLibsTest
Previously, SignatureMultiLibsTest extended SignatureTest but overrode
all the significant test methods so there was little point in extending
that class. The only test method (testCannotAccessUnexpectedClasses)
that was not overridden was a no-op in SignatureMultiLibsTest as it did
not have an unexpected-api-files argument.

Although they shared some information and both tested API signatures
they are otherwise very different. Separating them will allow each test
to be modified to make it more suited to their specific purpose.

This change separates the two, creates a common AbstractSignatureTest
base class for information that is shared and narrows the accessibility
of SignatureTest members. It also makes the expected-api-files and
previous-api-files required in SignatureMultiLibsTest.

Bug: 228322252
Test: cts/tests/signature/runSignatureTests.sh
      atest --host signature-host-tests
Change-Id: I90187eabe4cb37f7f6bfa696b2e9f6c2290a2a0a