Add CTS for not revoking permissions that have changed recently
Add a CTS test that takes an unused app, changes its permission, and
checks that the app's permissions are not auto-revoked afterward even
though the app is unused.

In order to do this in a reasonable time frame, we need to override a
setting to allow the storage to store the exact time instead of at the
granularity of a day.

Bug: 205642821
Test: atest PermissionChangeStorageImplTest
Change-Id: I4b3b07ee7fa07ca83ca051c9577ebb26bd68612b