bluetooth communication channel to connect two devices
once a device is associated, the app will establish a bluetooth
connection between the devices, so that data can be transferred.

The bluetooth communications channel setup is taken from a sample app
that sets up a client/server connection to send and receive data.

This CL also reorders the position of the buttons so that the "Start
Advertising" button is at the top since that's the first action to start
associating two devices

The basic flow is:
- App A starts advertising Device A -> App B requests to associate with Device A -> user accepts -> an association is created -> a bluetooth comms channel is established between the apps, so that data can be transferred.

- This is the same when App A associates with Device B. The only difference is that in the second case, the same comms channel is reused.(instead of creating a new one)


[Note that BLUETOOTH_CONNECT permission is required for the connection and advertising; so it is requested upfront. If it is not granted, the app is not functional]

Currently there is a sample message that is sent across. A followup CL will call startDataTransfer on CDM and fully implement the onMessageReceivedFromSystem method for the CompanionDeviceService

Demo - https://drive.google.com/file/d/11IGZanyrs8SjeUJK8GNJQDMYILnXg2iu/view?usp=sharing

BUG: 230382785
Change-Id: I259f0ef60f263430b6f662acac64da568fa042e9