DO NOT MERGE: fix CTS ImeInsetsVisibilityTest on Wear
This change disables swipe-to-dismiss for the dialog theme of the test
activity of ImeInsetsVisibilityTest on Wear. Enabling swipe-to-dismiss
on Wear has visibility implications during the test, in which case the
views are evaluated as occluding dynamically. This could cause the view
to be occluded when it should not. This is a similar change to
aosp/2028523

Skip-CP-To-TM: R specific changes

Bug: 232146121
Bug: 231923195

Test: run cts -m CtsInputMethodTestCases -t
android.view.inputmethod.cts.ImeInsetsVisibilityTest

Change-Id: I69468b3eae383ac28d5ede635eeb99e25ae4bf5c