Document model_arch_hash_64 in NNAPI host-side CTS test
A sibling CL in this topic adds model_arch_hash_64 to NNAPI platform
telemetry. This is a compressed 64-bit version of the existing
model_arch_hash field, which is 256 bits. model_arch_hash_64 is
compressed by bitwise-xor'ing all 64-bit chunks of model_arch_hash.

This CL adds the description to the CTS/MTS test that modelArchHash64
can have any value.

Bug: 232303482
Test: mma
Test: presubmit
Test: CtsNNAPIStatsdAtomHostTestCases
Change-Id: I9d43b249d9bbdd3e4014f9f8357949e4e621bd82