Exclude three tests in ActivityLifecycleTests
According to the tests flow first activity is opened and then a second
activity on another task. On Pixel is expects that the first
activity to be on stop since the second one is opened on top of it. On
Surface Duo the second activity is opened on the second task display area
instead of the same task display area as the first one. In this case the
expected assertion in the failing tests is no longer fulfilled.

Bug: 232962177

Tests: run cts -m CtsWindowManagerDeviceTestCases
-t
android.server.wm.lifecycle.ActivityLifecycleTests#testLocalRecreate

run cts -m CtsWindowManagerDeviceTestCases
-t
android.server.wm.lifecycle.ActivityLifecycleTests#testOnNewIntentFromHidden

run cts -m CtsWindowManagerDeviceTestCases
-t
android.server.wm.lifecycle.ActivityLifecycleTests#testTranslucentMovedIntoStack

Change-Id: Ib3183329eb86d1581e82f2b78911e5964274494f