Add assumption that task are in the the same TDA
According to the testing requirement, Assistant app shouldn't be visible
after AnimationTestActivity was launched. Because our device was built
with dual screen display area and by default design spec apps should be
only launched in one of the screen area. So, on Surface Duo the workflow
will be different. The test launches an AssistantActivity task in one
of the screens then the new AnimationTestActivity task launched with
Resume state on the other screen. At this point the condition
mWmState.waitAndAssertVisibilityGone (ASSISTANT_ACTIVITY) it is no
longer fulfilled because both tasks are resumed each on one pane, and
the visibility gone assertion for the AssistantActivity activity is
no longer fulfilled.

Bug: 232421385

Test: run cts -m CtsWindowManagerDeviceTestCases -t
android.server.wm.AssistantStackTests#testLaunchIntoSameTask

Change-Id: I4d6d201f5675290579ec653f45eece357e309d56