Specify launch TDA in TransitionSelectionTests
Some tests start multiple tasks and assume they will end up on the same
display area. This is not always the case for devices with multiple
screens.

Solution: Specifying the launch display area for CTS tasks allows OEMs
to meet the CTS assumptions. This approach launches consecutive tasks in
the same display area where the first task was launched.

Bug: 232962180
Test: cts-tradefed run commandAndExit -m CtsAppTestCases
		-t android.server.wm.TransitionSelectionTests
Verified On: Pixel 4 and development device.

Change-Id: I88ae67b46e4fd8cd85a85708d5a2c0dfd6568626