Mark HostsideConnOnActivityStartTest#testStartActivity_doze as flaky.
Marking the test as flaky while we work to identify the
root cause.

Also, for doze and appstandby related tests, we need
to enable the restriction everytime before starting
the activity. Add a TODO to update it.

Bug: 231440256
Test: atest tests/tests/app.usage/src/android/app/usage/cts/BroadcastResponseStatsTest.java
Change-Id: I97e8bdb6f3d0bd62bb98214ed0bf3cb0386bb21e
Merged-In: I97e8bdb6f3d0bd62bb98214ed0bf3cb0386bb21e