Fix the Cts testcase testSetForbiddenPlmnsTruncate.
Limiting the MAX_FPLMN_NUM to 100 in TelephonyManagerTest
results in failing of the testcase testSetForbiddenPlmnsTruncate
when the UICC supports more than 100 FPLMNs.
As per 3GPP TS 31.102 - 4.2.16 EFFPLMN (Forbidden PLMNs) spec,
there is no limitation to the number of FPLMNs that can be written
to the sim.

Increase the MAX_FPLMN_NUM to a very large value 1000 to  take
care of most cards.

Bug: 235017799
Test: TelephonyManagerTest#testSetForbiddenPlmnsTruncate
Change-Id: I7b43351c8c972f202ef82703fa0dea97ef47ec51