Add waitForIdle() after pressing home.
We've found in ag/18822380 that the home activity is launched when we
press the home key, so if we launch another activity immediately
afterwards it will be racy because the home activity launch might come
after our intended activity launch in some cases, which would result in
a test failure due to not finding the expected test activity. So add a
waitForIdle() after pressing home to remove that potential flakiness.

Test: presubmit
Change-Id: I7a3b717bd9ff3ec4dd81bad34b436e07f77e5456