Skip NintendoSwitchProTest
Cherry-pick rationale: b/285195363

Skip this test until we have the right set of responses figure out in
order to correctly register a virtual switch device via /dev/uhid.

Currently, when upstream driver version is used, the device registration
fails because we are not providing the right responses during the
initialization.

Bug: 227669886
Bug: 285195363
Test: m cts
Change-Id: I300543b44fd73e1a44f604a928ebb15716f7a385