Enforce DualSense support for keys and motions on all kernel versions
The joystick 'DualSense' can actually work with hid-generic, at least for the keys
and motions. Now that we support kernel driver dependent kls, we can enable this
test for all devices, not just the ones with the newer kernels.

For devices with the older kernels, the layout should fall back to the
non-hid-playstation one, and the correct keys should still be generated
as far as apps are concerned.

Bug: 228005926
Bug: 236864651
Test: atest CtsHardwareTestCases

Change-Id: I185b9b2a3b9a37be8800bf499383671651de6f45
Merged-In: I185b9b2a3b9a37be8800bf499383671651de6f45