Add missing test dependencies for CtsAppEnumerationTestCases
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependencies that are used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest-src CtsAppEnumerationTestCases \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsAppEnumerationTestCases
Merged-In: Ica6f734a8dc9284c8ca1ac4b8b673b42cf4e7a4f
Change-Id: Ie0988a2096edf03057f6dd7609166f020ab42d05