Add a missing test dependency for CtsExternalServiceTestCases
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependency that is used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest-src CtsExternalServiceTestCases \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsExternalServiceTestCases
Merged-In: I65711b08abb75a205770a037be94fd280dc104b9
Change-Id: I8ebc1f950eba4ee1b230fcfaa5cdc4e1de6c819e