Add missing test dependencies for CtsVoiceInteractionTestCases
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependencies that are used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest-src CtsVoiceInteractionTestCases \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsVoiceInteractionTestCases
Merged-In: I307369f7e3ac72b9cabab5c8aa635bcef009d737
Change-Id: I12ef360708ef123e811178a36f35920ba5bc1e09