Add a missing test dependency for CtsLibcoreOjTestCases
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependency that is used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest CtsLibcoreOjTestCases
Merged-In: Ica6f734a8dc9284c8ca1ac4b8b673b42cf4e7a4f
Change-Id: Iee29e22101f97f90288cfbc910434adca28444b6