Add `compatibility-host-provider-preconditions` for CtsProviderTestCases
The `compatibility-host-provider-preconditions` is required because the corresponding TF test config uses the `ExternalStoragePreparer` class which is defined in `compatibility-host-provider-preconditions`.

All of a test's dependencies should be explicitly declared in the Soong build module to enable infrastructure features that ensure correctness and optimize runtime. So that this change declares this dependency in the `host_required` attribute.

Bug: 220991354
Test: atest CtsProviderTestCases
Change-Id: I957be01ce541e92a3b8d929aec3467f56f55e4a9