Add duplicate classes from sdksandbox framework.jar to burndown list
The duplicate occurs within the sdksanbox framework.jar and sdksandbox
service.jar. The classes are within the same apex so it should be safe
to add them to the burndown list.

This is a temporary measure until b/242741880 is resolved and we are
able to jarjar properly. Until then, all AIDL files added to sdksandbox
framework.jar will have to be added to the burndown list.

This CL adds the existing duplicates.

Bug: 242741880
Test: atest StrictJavaPackagesTest
Change-Id: Iba7b9c0ba810404e0fdcbe44cbe513b0fc4d9e91