Rewrite InputDeviceKeyLayoutMapTest
Before this CL, the test was a no-op. It would inject keys through the
virtual input device, but it never checked whether the activity received
any of these keys.

It always passed, but also, the injected keys were still getting
processed by the system, so they would cause various activity launches.
These activities would come after the next test runs. They would cause
the next test to lose focus, and fail.

Some of the problems were:
1) KEYCODE_ prefix was getting appended twice
2) Errors were getting silently ignored
3) Check for whether a key should be skipped was faulty

In this CL, the test is rewritten to no longer ignore errors. Fragile
polling checks have been removed in favor of configuration change
monitoring. Conditions were corrected so that any error will lead to
failure. Tests were simplified so that the logic is easier to follow.

In addition, the meta checks are removed in this CL. These are out of
scope for this test. We can add a separate meta state test if desired.

Test: atest android.view.cts.input
Bug: 242085646
Change-Id: I404e38482be965b514dcd310f614ae57b6117263