Deflake hibnernation job completion assertions
Due to b/245797616, order of logcat statements was not actually being
enforced which made it possible for the test to continue before the job
actually completed.

We move off the logcat system entirely since it's somewhat brittle and
move to a (somewhat) less brittle check on the job state itself.

Bug: 241020365
Test: atest AutoRevokeTest --iterations 30 (with tests enabled)
Change-Id: I563d78561d073d7247c640fceaae0f97448a2a52