Fix LauncherAppsEnumerationTests#callback_unsuspended_visibleReceives
With Ie06df12ff3da6cf209fb3149e6aa80b0e63ce56e, broadcasts are not
sent if the package being unsuspended didn't have a single
suspending package. And this test was written when the original bug
was introduced, so it tests the wrong behavior.

This makes the test suspend the package first before testing the
unsuspend.

Bug: 232319688

Test: atest LauncherAppsEnumerationTests

Change-Id: I3275a06e9c3c95acac413cb485250333ea3548fb