Fix test flaky in ImeInsetsVisibilityTest
As in ImeInsetsVisibilityTest uses emulateTapOnViewCenter to simulate
tapping the editor view to show IME, it may flaky when the tapping even
too early before the test activity fully drawn because of the animation.

Uses View#requestFocus + showSoftInput to fix this unnecessary flakiness
since originally the test is not required to test tapping focus
behavior.

Fix: 243196094
Test: atest ImeInsetsVisibilityTest --rerun-until-failure 10
Change-Id: I428539bebb02fe348c23b313293f3d7b1768c1cd