CodecEncoderSurfaceTest: Skip tests if decoder doesn't support P010
CodecEncoderSurfaceTest sets decoder's color format as P010 when
testing encoders that support 10-bit encoding.
It currently picks decoder by using MediaCodecList's
findDecoderForFormat(), but this method ignores color format and
returns a decoder that on some devices may not support P010.
In such devices where decoder doesn't support P010, the test is
now skipped with an assumption failure in the report.

Bug: 241022151
Test: atest CtsMediaV2TestCases:CodecEncoderSurfaceTest

Change-Id: I20094582a06acdacfe3dd33a1cd00315321964c0