CameraITS: Use a fallback recording bitrate instead of crashing
The PreviewRecorder queries CamcorderProfile to get an upper limit of
supported bitrate for a given size and camera. However, since
CamcorderProfiles aren't exhaustive, the test can run into situations
where it fails to find a supported upper limit. This situation only
arises for higher resolutions for cameras that support a maximum
resolution between QUALITY_1080P and QUALITY_2K.

The CL adds a fallback constant bitrate to be used when this situation
occurs.

Bug: 248393376
Bug: 244673213
Test: Manually tested that the fallback value is used
Change-Id: I094b989ed349d4de358f3c44191fcd053ac1cea9