Fix testHotwordDetectionService_onDetectFromDsp_rejection fail
After this change(ag/20034767) is merged, it seems that we
meet the startInput permission denied in AudioFlinger when
starting audio record. It seems that we need to use
uiAutomation.adoptShellPermissionIdentity to retain the
identity until the callback is triggered.

This fix will not break the public release and pass the fail
test after ag/20034767.

Bug: 249314302
Test: atest HotwordDetectionServiceBasicTest --iterations
Change-Id: I0b41b346f3b0844f8f35a00916e4e52cc058e399