Wait for the Wifi disabling completed before checking the network status.
In the CTS test testGetCurrentNetwork after disabling Wifi, add polling
check to make sure Wifi is disabled.
The Wifi network connection info is cleared when the supplcante state
changed to disconnect, while the current network is set to updated when
the client mode exits ConnectableState.

Fixed the flaky test with possible failure when network connection info
is cleared but state machine not exit.

Bug: 252814919
Test: atest android.net.wifi.cts.WifiManagerTest#testGetCurrentNetwork
Change-Id: Ief5c366f1b2fd43e94940c707309ec0c690fa104