Prevent .quals() being invoked on null object
Currently mLockType becomes null when any of the clear methods are called. This
means that if the lock type is queried after (using hasPassword,
hasPattern, hasPin, password, pattern or pin) .quals is invoked on a
null object.

To fix this, the object invoking .quals() and the object passed are swapped
so the lock type (which cannot be null) invokes .equals and mLockType
(which can be null) is the object passed.

Bug: 253073550
Test: atest UserReferenceTest
Ignore-AOSP-First: No confidential information
Change-Id: Iffd3bc3942fc203b04b2df41b7c55572e0142568