Fix wrong userId being reported to CUJ metrics
CUJ does not need to know user id before it is created as long as begin and end session is the same.
User Journey Create User Begin will report user id -1 as it was not yet created.
On finish it will semd corrent user id or -1 if user creation faild.

Bug: 251357485
Bug: 251357372

Test: atest CtsStatsdAtomHostTestCases:android.cts.statsdatom.users.UserLifecycleEventStatsTests#testCreateFullUser
Test: atest CtsStatsdAtomHostTestCases:android.cts.statsdatom.users.UserLifecycleEventStatsTests#testCreateGuestUser
Change-Id: I115e9019e9dcd7a00bdf167305bec7eb48bce570