Revert "Add test for report[Mobile|Wifi]RadioPowerState APIs"
This reverts commit ad15d78d2462c8f3423cf6a34c929ab013215998.

The available way to verify report[Mobile|Wifi]RadioPowerState is
to verify what is dumped in the dumpsys. These calls are intended to notify cellular/wifi network switch stats. If BatteryStats is no
longer actively support dumpsys, there is no way to verify the result
and is less useful to verify this.

Reason for revert: <no longer actively support dumpsys>
Fix: 253767052
Change-Id: I1f4e577796b41463f5560508d4863b8e57ad9c89