Refactor AlwaysOnHotwordDetectorTest(2/2)
Currently, the voice cts communication across three packages. It
makes the test hard to read and we need to use broadcast, service
to let the data can pass between these thress packages.

To reduce test complexability, we will try to make the test
communication in the same test package that can access service
variables or methods easier.

The BasicVoiceInteractionService is used by many test cases, in this
change, it only moves the logic used by AlwaysOnHotwordDetectorTest
to CtsBasicVoiceInteractionService. The test can access the service
instance directly now, we don't need to the intermediate Activities,
services. We will remove the out package BasicVoiceInteractionService

Bug: 249287756
Test: atest CtsVoiceInteractionTestCases

Change-Id: I893c6c488e24b3f71bcb120eedfdccd8a0b33b00