Add delay when turning OFF BluetoothAdapter.
On Cuttlefish devices, there is a delay in the events
when turning OFF Bluetooth. This causes an event turning
OFF to appear when Bluetooth is turned ON immediately
after being turned OFF.
Adding a delay in the tests after turning OFF Bt is a
temporary fix that will be removed once the related bug
has been fixed.

Bug: 234892968
Tag: #feature
Test: atest CtsBluetoothTestCases
Change-Id: I5d878d14b806be50ad8d8ffef2f5638c919133eb