Move hibernation tests to its own module
Create a new module for hibernation tests so that we can more
granularly control how its run in presubmit/postsubmit and how the test
is set up.

This CL should do nothing else besides the functional changes required
for the refactor in order to limit the possible causes of things
breaking through this set of changes.

Bug: 248087707
Test: atest CtsHibernationTestCases
Test: atest CtsOsTestCases
Change-Id: I0fcd36fe2f7370806272633680d31aa2819c0668