Fix NotificationListenerVerifierActivity[folded] no result
From Android T, the "Notification Permission" is a dangerous permission.
After executing "Please block the linke application and return here",
the "Notification Permission" will be revoked, and the
"PermissionManagerService" will kill the CV process.

When returning to "NotificationListenerVerifierActivity", the CV process
restarts, but not execute "TestListActivity.createContinue()", so
"sCurrentDisplayMode" is the default value of "unfolded". Finally, the
result of NotificationListenerVerifierActivity[folded] is 0 (not
executed).

Bug:255265824

Test: CTS VERIFIER[Folded]-Notification Listener Test

Signed-off-by: duanyangqi <duanyangqi@xiaomi.com>
Change-Id: I5147fb8ad5094811a223a155c2d0c2cbfc39d364