Allow longer delays for peeking videos and provide frame rate
Some devices have display pipelines with high latencies that don't
allow for peeking in a timely manner.

Some devices require the frame rate (either explicit or inferred) in
order to configure the display hardware. In the future, this hardware
should be updated to accommodate rendering a video frame without
configuring the display hardware with a frame rate.

Bug: 199394267
Fixes: 199394267
Test: atest android.media.decoder.cts.DecoderTest#testTunneledVideoPeekDefault
Change-Id: I0ea89f3e8284198981965e7f8e16a9cda196e5ea