webkit/sdksandbox: Use new support for before/after
Now that the test library we use to execute
tests within the SDK sandbox supports Before and After
annotations, we can clean up our tests since we know
the setup and clean up methods will also run
in the sandbox.

This makes our tests a lot clearer.

We will still need to keep a few member variables around
for a bit longer until we are referencing the test
environment in all our tests.

Test: atest CtsSdkSandboxWebkitTestCases && atest CtsWebkitTestCases:WebViewTest
Bug: 259548800
Change-Id: I6b114073815e5459ed436dfa5c22b87d7b90c29a