Don't wait for broadcast barrier by default.
Currently, BlockingBroadcastReceiver.awaitForBroadcast()
variants always call wait-for-broadcast-barrier which
could take a long time to finish in some situations where
there are BOOT_COMPLETED or PACKAGE_CHANGED (or other
broadcasts that are subscribed to by a lot of receivers)
broadcasts sent and cause the tests to fail. So, let tests
invoke the wait-for-broadcast-barrier explicitly if they
need it and have these tests marked as @LargeTest.

Bug: 260158381
Test: TH
Change-Id: I55751cb480154675bdeb37a0c7e0ba0cc6bcf194