Add Logging to Carrier Privileged Test to Debug Future Failures
There was a small blip of intermittent failures in
testSetAllowedNetworkTypesForReason_carrierPrivileges. This change
splits the handling of the behavior under test (SecurityException) from
the required handling of the checked Exception by the test carrier
privilege test utility.

Failed tests were not reproducible locally, so the hope is that this
extra logging provides more information if this happens again.

Bug: 262969190
Test: atest "CtsTelephonyTestCases:android.telephony.cts.TelephonyManagerTest#testSetAllowedNetworkTypesForReason_carrierPrivileges"
Change-Id: I0e55c2fec1c638f2da937a0651e8bb72787371d2