Fix flakiness in DisplayWakeReportedStatsTests
These tests wake the device in various ways.
But first, the device must be put to sleep.
In some cases, the wake-action was executed while the device was still
transitioning to the Asleep state (in the Dozing state).
With this change we wait for the device to enter the Asleep state before
continuing.
It also adds the showWhenLocked attribute to the activity, so the screen
can also wake in case there is a lockscreen showing.

Bug: 263749497
Test: atest --iterations 10 DisplayWakeReportedStatsTests
Change-Id: I23a4f9802c8074e6b5be9d8d06fec6bd46f59851