Add tests to verify rotation of compromised key in sharedUserId
To facilitate rotation of a compromised key in a sharedUserId, the
platform will allow an app that is already part of a sharedUserId
to be updated when it has rotated its signing key and revoked the
SHARED_USER_ID capability of the previous key even if other
packages in the sharedUserId are still signed by the untrusted key.
This commit adds / updates tests to verify an update is supported
in this scenario but an install is prevented until all apps in the
sharedUserId have been rotated off the untrusted key.

Bug: 243701925
Test: atest PkgInstallSignatureVerificationTest
Change-Id: I611b1f1574944b54094a96a09340e2f0edaeefbd