Add queries tag for V3RotationTest to query test package
The PkgInstallSignatureVerificationTest hostside test uses the
V3RotationTest device side class to run a number of rotation tests
against separate test apps installed on the device. All test apps
installed as part of CTS should currently have the force-queryable
flag set to grant visibility to other test apps, but if this is ever
removed, the rotation tests would fail due to the device class not
being able to query the signatures of the test app. This commit adds
queries tags to the V3RotationTest app to ensure it can query the
signatures of the test and companion apps.

Fixes: 262822609
Test: atest PkgInstallSignatureVerificationTest
Change-Id: I5603af7ddc847f7858e2d82f94edf8500ccdf55a