Add exported flag to exposed test runtime receivers
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting U
and registering a receiver for non-system broadcasts must
specify either the exported or not exported flag when invoking
registerReceiver; if one of these flags is not provided, the
platform will throw a SecurityException. This commit updates
the receivers in MyInteractionService with the RECEIVER_EXPORTED
flag to maintain the current behavior of the test.

Fixes: 264551884
Test: atest MixedDeviceOwnerTest#testScreenCaptureDisabled_assist
Change-Id: Idea6175712f493d48bb23f6d22773edf32225174