Updates wait logic in AccessibilityOverlayTest.
This test was flaky (or outright failing on some devices) because the
virtual display is not guaranteed to have a default window, so it was
not guaranteed to be visible to accessibility.

This replaces the faulty wait logic in
testA11yServiceShowsOverlayOnVirtualDisplay_shouldAppear
with adding a default activity window, which forces the virtual display
to be recognized and tracked by AccessibilityWindowManager.

Bug: 263225292
Bug: 282832881
Test: atest AccessibilityOverlayTest#testA11yServiceShowsOverlayOnVirtualDisplay_shouldAppear --iterations 100
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:58fa46c937c09464c550dbb094ddd7cb69d60da3)

Merged-In: I273dddff0e207e75b8ba56a517618d8481e99886
Change-Id: I273dddff0e207e75b8ba56a517618d8481e99886