Complete injected gestures
The injected gestures must be consistent and complete.
Currently, the test is not injecting ACTION_UP events for some of the
gestures, which causes the system to (correctly) reject the other
gestures that are trying to be injected.

Since the globally injected gestures are supposed to emulate real
devices, all of the injected events must be self-consistent.

Bug: 266183437
Test: atest android.view.inputmethod.cts.StylusHandwritingTest
Change-Id: I4793030d1221f7fc00146cd391da4f9dac76e168