When navigation bar is towards the left side, 4 of the test cases are failing from CtsWindowManagerDeviceTestCases.
Test cases try to position two views one over the other and test
the obscure views functionality.
The 4 tests are failing as the views are not placed one over
the other in the softwares where navigation bar is on Left side.
As the tests are related to functionality of obscure views, to avoid
this failure in different system bar configurations,
the content is made to be full screen. So that the functionality of
obscure views can be tested in all system bar configurations.

Bug: 263295540
Test: Executed CtsWindowManagerDeviceTestCases module android.server.wm.WindowInputTests test class
Change-Id: Iec7c84c99170fbcc552c32c7137840bb6152dd39