Always clear signal strength update request to remove flakiness
This CL makes sure that in TelephonyManagerTest CTS cases, the
clearSignalStrengthUpdateRequest is always called after
setSignalStrengthUpdateRequest to remove the flakiness.

API setSignalStrengthUpdateRequest is only allow to call once
with same subId in a process. Without clearing request, the
previous setting request may impact the following test cases.

clearSignalStrengthUpdateRequest is an no-op if the input request
was not set (and it is non-null). So clearing call will have no
side effort if the previous setting operation failed.

Bug: 266832527
Test: atest android.telephony.cts.TelephonyManagerTest
Change-Id: Ib9a40a41c1dd69c4529760bb15ccf1115be84c66