Enabled stricter DualSense Edge vibration checks
When DualSense Edge CTS was first commited, the x86 prebuilt kernel that
the build test system was using was out of date and not allowing tests
to pass properly. Vibration tests were relaxed to allow it to pass. Once
the prebuilt was updated we could return the tests to their original
more strict form.

Bug: 260685629
Test: atest SonyDualSenseEdgeBluetoothTest SonyDualSenseEdgeUsbTest
Change-Id: I2e585268ff110b547267e28d409f7a8bed006c47