Fix CTS tests timing out
This reverts ag/21271246 which was incorrectly merged because tests were
not properly verified. Move DeviceSetup preparation to before the APK
installations to make sure that settings are set before attempting to
install any APKs.
Also reduce the timeout for sending MMS tests from 5m->30s. MMS is not supported on CF, so this test will always fail. Temporarily reduce the timeout until a solution can be found to prevent the entire test suite from timing out.

Test: atest CtsTelephonyTestCases from clean build
Bug: 224792744
Change-Id: Ia6d16a9806afc702dc819e63d757e04b4dc45859