webkit: WebView test readability improvement
Add a method to SharedWebViewTestEnvironment that returns a new
test web server that has also been started. Update test methods to
call this method instead of separate methods for getting the web server
and then starting it. This removes the need in a lot of test classes to
have a helper function that performs these two tasks. Each test now also
explicitly states the sslmode it needs the web server to use making it
easier to read.

Test: atest CtsWebkitTestCases
Change-Id: Ie6b17033834a92a4a721c75d7072ffd816dad765