Fix BackNavigationTests flaky
The `uiautomation.waitForIdle` is based on the a11y event, which would
potentially cause the timeout if the a11y manager service didn't update
the info on time.

This CL will replace it with `Instrumentation.waitForIdleSync` to ensure
the application is idle and prevent the unnecessary timeout checking.

Also wake and unlock device before test start, that could prevent test
failed when screen off.

Bug: 263018398
Test: atest BackNavigationTests BackNavigationLegacyTest --rerun-until-failure 100
Change-Id: I803e294d5cc6a9c98d0012d3e02ff5d5f3b34f5c