Dismiss dialogs in clipboard tests.
The clipboard related tests are flaky because focus is required
in order to access clipboiard, and sometimes in automation dialogs
appear over the test activity, causing focus to be lost.

This CL adds home and back presses to the start of every test to
clear dialogs before the tests are run.

Bug: 267616703
Bug: 267616842
Bug: 267617210
Bug: 266033004
Bug: 265250261
Bug: 265250261
Test: atest ClipboardManagerTest
Test: atest ClipboardManagerListenerTest
Test: atest ClipDescriptionTest
Change-Id: I23bfed1f9e271701fc0d598ddf1b385edf6d8595