Allow tagging CTS tests to be run on older platform versions
Currently, the min SDK level gets inherited from the place where the build happens. Now that UDC is snapped off, this means that building CTS on master, and then trying to run it on UDC, means that the test device's SDK level is too low to run the test.

Patch up a few areas where the min SDK version is wrong, or is
implicitly figured out wrong, so the tests can run successfully.

Bug: 270704752
Change-Id: Ia2c7480ef4a4c187300b390662ed658856c1542c
Test: CtsTaggingHostTestCases from git_master on UDC device.