Make sure to leave the screen on at end of KeepScreenOnTests
Other tests that run after KeepScreenOnTests may assume the screen is
already turned on (the atest infrastructure turns it on before
beginning to run tests), but some of the KeepScreenOnTests
purposefully cause the screen to turn off. This CL makes sure to turn
it back on on at the end of these tests.

Bug: 270993328
Test: atest KeepScreenOnTests (and any other test afterwards that needs screen on)
Change-Id: I55f1cf2cc969735a1907098ddc897949b4e47626