Avoid changing orientation when running EnsureBarContrastTest
The testing device may be put with non natural rotation with
auto rotation enabled. Then the test may capture an intermediate
state when the activity is transforming. The visual result is no
problem but the snapshot captured by the test may use the bounds
in different orientation. So make the test activity keep the
current orientation to stabilize the test.

Bug: 268167570
Test: atest EnsureBarContrastTest
Change-Id: I081d2ac2847d66796c43b35de8b47b312d0dda70