Update the test to verify the activity state
If DEFAULT_ENABLE_WAIT_FOR_FINISH_ATTACH_APPLICATION is true, the
activity process has to wait for the debugger to be attached
and won't notify ATM before that. So, the activity will be stayed
in INITIALIZING state and unable to get the process id from WM proto.

Also fix a typo that slows down the test for 10 seconds.

Bug: 271477038
Test: AmStartOptionsTests
Change-Id: Ib25e58b14825655c1c4450f8b067786acb3ec589