Claim ownership of the CTS test server for Cronet.
The CTS test server is currently abandoned and ownership rolls up to
owners of the cts/ package who understandably don't have the bandwidth
of reviewing new additions. Cronet is one of the largest users of the
test server, and also owns the Android platform HTTP client.

The CL is being sent to btmura@ who originally added the test server to
AOSP (471beda), and torne@ who was the last person to send a feature CL
(c77e115).

Test: N/A
Change-Id: I9e36125aee9006af89f700fc35e94eb08b563dfc