Disable WiredHeadsetTests in Telecom CTS.
Due to changes made as part of this bug, the media session which used to
be opened whenever there is a call is now only opened when a wired headset
is attached.  Given the purpose of the media session is to enable control
of calls via wireless headsets, this change makes sense.

However, these tests can no longer pass because a CTS test cannot route
the call to a wireless headset if none is attached to the device.
I've created a followup bug to see if we can make a way to fake out audio
routes in CTS so we could re-enable this CTS.  That is a pretty big change
so for now we will just disable these tests.

Test: Run CTS tests and verify no longer failing.
Fixes: 266171157
Change-Id: Ie122a505f9af9a1b0be4934024f887aaae340e7d