Revert "Camera CTS: Fix isNoiseReductionModeControlSupported"
This reverts commit 14e4f7e4387212dbbbd93f7d771e1bdc802442d6.

Reason for revert: logic error in checking noiseReductionMode
not being supported.

Per HAL spec:
HAL must support both FAST and HIGH_QUALITY if noise reduction
control is available on the camera device, but the underlying
implementation can be the same for both modes. That is, if the
highest quality implementation on the camera device does not
slow down capture rate, then FAST and HIGH_QUALITY will generate
the same output.

Bug: 272479897
Test: vendor test
Change-Id: Ic65d42c650efef0a03abb38dd9cf6cc66b59e380